Crate aws_sdk_dynamodb

Amazon DynamoDB

Amazon DynamoDB is a fully managed NoSQL database service that provides fast and predictable performance with seamless scalability. DynamoDB lets you offload the administrative burdens of operating and scaling a distributed database, so that you don't have to worry about hardware provisioning, setup and configuration, replication, software patching, or cluster scaling.

With DynamoDB, you can create database tables that can store and retrieve any amount of data, and serve any level of request traffic. You can scale up or scale down your tables' throughput capacity without downtime or performance degradation, and use the AWS Management Console to monitor resource utilization and performance metrics.

DynamoDB automatically spreads the data and traffic for your tables over a sufficient number of servers to handle your throughput and storage requirements, while maintaining consistent and fast performance. All of your data is stored on solid state disks (SSDs) and automatically replicated across multiple Availability Zones in an AWS region, providing built-in high availability and data durability.

Re-exports

pub use config::Config;

pub use client::Client;

Modules

client

config

error

input

model

operation

output

Structs

Blob

ByteStream

Stream of binary data

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

Enums

Error

SdkError

Failing Sdk Result

Statics

PKG_VERSION