Module aws_sdk_dynamodb::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

BatchExecuteStatement

Fluent builder constructing a request to BatchExecuteStatement.

BatchGetItem

Fluent builder constructing a request to BatchGetItem.

BatchWriteItem

Fluent builder constructing a request to BatchWriteItem.

CreateBackup

Fluent builder constructing a request to CreateBackup.

CreateGlobalTable

Fluent builder constructing a request to CreateGlobalTable.

CreateTable

Fluent builder constructing a request to CreateTable.

DeleteBackup

Fluent builder constructing a request to DeleteBackup.

DeleteItem

Fluent builder constructing a request to DeleteItem.

DeleteTable

Fluent builder constructing a request to DeleteTable.

DescribeBackup

Fluent builder constructing a request to DescribeBackup.

DescribeContinuousBackups

Fluent builder constructing a request to DescribeContinuousBackups.

DescribeContributorInsights

Fluent builder constructing a request to DescribeContributorInsights.

DescribeEndpoints

Fluent builder constructing a request to DescribeEndpoints.

DescribeExport

Fluent builder constructing a request to DescribeExport.

DescribeGlobalTable

Fluent builder constructing a request to DescribeGlobalTable.

DescribeGlobalTableSettings

Fluent builder constructing a request to DescribeGlobalTableSettings.

DescribeKinesisStreamingDestination

Fluent builder constructing a request to DescribeKinesisStreamingDestination.

DescribeLimits

Fluent builder constructing a request to DescribeLimits.

DescribeTable

Fluent builder constructing a request to DescribeTable.

DescribeTableReplicaAutoScaling

Fluent builder constructing a request to DescribeTableReplicaAutoScaling.

DescribeTimeToLive

Fluent builder constructing a request to DescribeTimeToLive.

DisableKinesisStreamingDestination

Fluent builder constructing a request to DisableKinesisStreamingDestination.

EnableKinesisStreamingDestination

Fluent builder constructing a request to EnableKinesisStreamingDestination.

ExecuteStatement

Fluent builder constructing a request to ExecuteStatement.

ExecuteTransaction

Fluent builder constructing a request to ExecuteTransaction.

ExportTableToPointInTime

Fluent builder constructing a request to ExportTableToPointInTime.

GetItem

Fluent builder constructing a request to GetItem.

ListBackups

Fluent builder constructing a request to ListBackups.

ListContributorInsights

Fluent builder constructing a request to ListContributorInsights.

ListExports

Fluent builder constructing a request to ListExports.

ListGlobalTables

Fluent builder constructing a request to ListGlobalTables.

ListTables

Fluent builder constructing a request to ListTables.

ListTagsOfResource

Fluent builder constructing a request to ListTagsOfResource.

PutItem

Fluent builder constructing a request to PutItem.

Query

Fluent builder constructing a request to Query.

RestoreTableFromBackup

Fluent builder constructing a request to RestoreTableFromBackup.

RestoreTableToPointInTime

Fluent builder constructing a request to RestoreTableToPointInTime.

Scan

Fluent builder constructing a request to Scan.

TagResource

Fluent builder constructing a request to TagResource.

TransactGetItems

Fluent builder constructing a request to TransactGetItems.

TransactWriteItems

Fluent builder constructing a request to TransactWriteItems.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateContinuousBackups

Fluent builder constructing a request to UpdateContinuousBackups.

UpdateContributorInsights

Fluent builder constructing a request to UpdateContributorInsights.

UpdateGlobalTable

Fluent builder constructing a request to UpdateGlobalTable.

UpdateGlobalTableSettings

Fluent builder constructing a request to UpdateGlobalTableSettings.

UpdateItem

Fluent builder constructing a request to UpdateItem.

UpdateTable

Fluent builder constructing a request to UpdateTable.

UpdateTableReplicaAutoScaling

Fluent builder constructing a request to UpdateTableReplicaAutoScaling.

UpdateTimeToLive

Fluent builder constructing a request to UpdateTimeToLive.