Struct aws_sdk_dynamodb::client::fluent_builders::DeleteItem

pub struct DeleteItem<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> DeleteItem<C>

pub async fn send(self) -> Result<DeleteItemOutput, SdkError<DeleteItemError>> where
C: SmithyConnector,

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table from which to delete the item.

pub fn set_table_name(self, input: Option<String>) -> Self

pub fn key(self, k: impl Into<String>, v: impl Into<AttributeValue>) -> Self

A map of attribute names to AttributeValue objects, representing the primary key of the item to delete.

For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

pub fn set_key(self, input: Option<HashMap<String, AttributeValue>>) -> Self

pub fn expected(
self,
k: impl Into<String>,
v: impl Into<ExpectedAttributeValue>
) -> Self

This is a legacy parameter. Use ConditionExpression instead. For more information, see Expected in the Amazon DynamoDB Developer Guide.

pub fn set_expected(
self,
input: Option<HashMap<String, ExpectedAttributeValue>>
) -> Self

pub fn conditional_operator(self, input: ConditionalOperator) -> Self

This is a legacy parameter. Use ConditionExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide.

pub fn set_conditional_operator(
self,
input: Option<ConditionalOperator>
) -> Self

pub fn return_values(self, input: ReturnValue) -> Self

Use ReturnValues if you want to get the item attributes as they appeared before they were deleted. For DeleteItem, the valid values are:

NONE - If ReturnValues is not specified, or if its value is NONE, then nothing is returned. (This setting is the default for ReturnValues.)

ALL_OLD - The content of the old item is returned.

The ReturnValues parameter is used by several DynamoDB operations; however, DeleteItem does not recognize any values other than NONE or ALL_OLD.

pub fn set_return_values(self, input: Option<ReturnValue>) -> Self

pub fn return_consumed_capacity(self, input: ReturnConsumedCapacity) -> Self

Determines the level of detail about provisioned throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn set_return_consumed_capacity(
self,
input: Option<ReturnConsumedCapacity>
) -> Self

pub fn return_item_collection_metrics(
self,
input: ReturnItemCollectionMetrics
) -> Self

Determines whether item collection metrics are returned. If set to SIZE, the response includes statistics about item collections, if any, that were modified during the operation are returned in the response. If set to NONE (the default), no statistics are returned.

pub fn set_return_item_collection_metrics(
self,
input: Option<ReturnItemCollectionMetrics>
) -> Self

pub fn condition_expression(self, input: impl Into<String>) -> Self

A condition that must be satisfied in order for a conditional DeleteItem to succeed.

An expression can contain any of the following:

Functions: attribute_exists | attribute_not_exists | attribute_type | contains | begins_with | size

These function names are case-sensitive.

Comparison operators: = | <> | < | > | <= | >= | BETWEEN | IN

Logical operators: AND | OR | NOT

For more information about condition expressions, see Condition Expressions in the Amazon DynamoDB Developer Guide.

pub fn set_condition_expression(self, input: Option<String>) -> Self

pub fn expression_attribute_names(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see Specifying Item Attributes in the Amazon DynamoDB Developer Guide.

pub fn set_expression_attribute_names(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn expression_attribute_values(
self,
k: impl Into<String>,
v: impl Into<AttributeValue>
) -> Self

One or more values that can be substituted in an expression.

Use the : (colon) character in an expression to dereference an attribute value. For example, suppose that you wanted to check whether the value of the ProductStatus attribute was one of the following:

Available | Backordered | Discontinued

You would first need to specify ExpressionAttributeValues as follows:

{ ":avail":{"S":"Available"}, ":back":{"S":"Backordered"}, ":disc":{"S":"Discontinued"} }

You could then use these values in an expression, such as this:

ProductStatus IN (:avail, :back, :disc)

For more information on expression attribute values, see Condition Expressions in the Amazon DynamoDB Developer Guide.

pub fn set_expression_attribute_values(
self,
input: Option<HashMap<String, AttributeValue>>
) -> Self