Struct aws_sdk_dynamodb::client::fluent_builders::ExecuteStatement

pub struct ExecuteStatement { /* private fields */ }

Fluent builder constructing a request to ExecuteStatement.

This operation allows you to perform reads and singleton writes on data stored in DynamoDB, using PartiQL.

For PartiQL reads (SELECT statement), if the total number of processed items exceeds the maximum dataset size limit of 1 MB, the read stops and results are returned to the user as a LastEvaluatedKey value to continue the read in a subsequent operation. If the filter criteria in WHERE clause does not match any data, the read will return an empty result set.

A single SELECT statement response can return up to the maximum number of items (if using the Limit parameter) or a maximum of 1 MB of data (and then apply any filtering to the results using WHERE clause). If LastEvaluatedKey is present in the response, you need to paginate the result set.

Implementations

impl ExecuteStatement

pub async fn send( self ) -> Result<ExecuteStatementOutput, SdkError<ExecuteStatementError>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn statement(self, input: impl Into<String>) -> Self

The PartiQL statement representing the operation to run.

pub fn set_statement(self, input: Option<String>) -> Self

The PartiQL statement representing the operation to run.

pub fn parameters(self, input: AttributeValue) -> Self

Appends an item to Parameters.

To override the contents of this collection use set_parameters.

The parameters for the PartiQL statement, if any.

pub fn set_parameters(self, input: Option<Vec<AttributeValue>>) -> Self

The parameters for the PartiQL statement, if any.

pub fn consistent_read(self, input: bool) -> Self

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

pub fn set_consistent_read(self, input: Option<bool>) -> Self

The consistency of a read operation. If set to true, then a strongly consistent read is used; otherwise, an eventually consistent read is used.

pub fn next_token(self, input: impl Into<String>) -> Self

Set this value to get remaining results, if NextToken was returned in the statement response.

pub fn set_next_token(self, input: Option<String>) -> Self

Set this value to get remaining results, if NextToken was returned in the statement response.

pub fn return_consumed_capacity(self, input: ReturnConsumedCapacity) -> Self

Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn set_return_consumed_capacity( self, input: Option<ReturnConsumedCapacity> ) -> Self

Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn limit(self, input: i32) -> Self

The maximum number of items to evaluate (not necessarily the number of matching items). If DynamoDB processes the number of items up to the limit while processing the results, it stops the operation and returns the matching values up to that point, along with a key in LastEvaluatedKey to apply in a subsequent operation so you can pick up where you left off. Also, if the processed dataset size exceeds 1 MB before DynamoDB reaches this limit, it stops the operation and returns the matching values up to the limit, and a key in LastEvaluatedKey to apply in a subsequent operation to continue the operation.

pub fn set_limit(self, input: Option<i32>) -> Self

The maximum number of items to evaluate (not necessarily the number of matching items). If DynamoDB processes the number of items up to the limit while processing the results, it stops the operation and returns the matching values up to that point, along with a key in LastEvaluatedKey to apply in a subsequent operation so you can pick up where you left off. Also, if the processed dataset size exceeds 1 MB before DynamoDB reaches this limit, it stops the operation and returns the matching values up to the limit, and a key in LastEvaluatedKey to apply in a subsequent operation to continue the operation.