Struct aws_sdk_dynamodb::client::fluent_builders::GetItem

pub struct GetItem<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> GetItem<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<GetItemOutput, SdkError<GetItemError>> where
R::Policy: SmithyRetryPolicy<GetItemInputOperationOutputAlias, GetItemOutput, GetItemError, GetItemInputOperationRetryAlias>,

pub fn table_name(self, inp: impl Into<String>) -> Self

The name of the table containing the requested item.

pub fn set_table_name(self, input: Option<String>) -> Self

pub fn key(self, k: impl Into<String>, v: impl Into<AttributeValue>) -> Self

Adds a key-value pair to Key.

To override the contents of this collection use set_key.

A map of attribute names to AttributeValue objects, representing the primary key of the item to retrieve.

For the primary key, you must provide all of the attributes. For example, with a simple primary key, you only need to provide a value for the partition key. For a composite primary key, you must provide values for both the partition key and the sort key.

pub fn set_key(self, input: Option<HashMap<String, AttributeValue>>) -> Self

pub fn attributes_to_get(self, inp: impl Into<String>) -> Self

Appends an item to AttributesToGet.

To override the contents of this collection use set_attributes_to_get.

This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide.

pub fn set_attributes_to_get(self, input: Option<Vec<String>>) -> Self

pub fn consistent_read(self, inp: bool) -> Self

Determines the read consistency model: If set to true, then the operation uses strongly consistent reads; otherwise, the operation uses eventually consistent reads.

pub fn set_consistent_read(self, input: Option<bool>) -> Self

pub fn return_consumed_capacity(self, inp: ReturnConsumedCapacity) -> Self

Determines the level of detail about provisioned throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn set_return_consumed_capacity(
self,
input: Option<ReturnConsumedCapacity>
) -> Self

pub fn projection_expression(self, inp: impl Into<String>) -> Self

A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes are returned. If any of the requested attributes are not found, they do not appear in the result.

For more information, see Specifying Item Attributes in the Amazon DynamoDB Developer Guide.

pub fn set_projection_expression(self, input: Option<String>) -> Self

pub fn expression_attribute_names(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to ExpressionAttributeNames.

To override the contents of this collection use set_expression_attribute_names.

One or more substitution tokens for attribute names in an expression. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information on expression attribute names, see Specifying Item Attributes in the Amazon DynamoDB Developer Guide.

pub fn set_expression_attribute_names(
self,
input: Option<HashMap<String, String>>
) -> Self