Struct aws_sdk_dynamodb::client::fluent_builders::ListBackups

pub struct ListBackups { /* private fields */ }

Fluent builder constructing a request to ListBackups.

List backups associated with an Amazon Web Services account. To list backups for a given table, specify TableName. ListBackups returns a paginated list of results with at most 1 MB worth of items in a page. You can also specify a maximum number of entries to be returned in a page.

In the request, start time is inclusive, but end time is exclusive. Note that these boundaries are for the time at which the original backup was requested.

You can call ListBackups a maximum of five times per second.

Implementations

impl ListBackups

pub async fn send(self) -> Result<ListBackupsOutput, SdkError<ListBackupsError>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn table_name(self, input: impl Into<String>) -> Self

The backups from the table specified by TableName are listed.

pub fn set_table_name(self, input: Option<String>) -> Self

The backups from the table specified by TableName are listed.

pub fn limit(self, input: i32) -> Self

Maximum number of backups to return at once.

pub fn set_limit(self, input: Option<i32>) -> Self

Maximum number of backups to return at once.

pub fn time_range_lower_bound(self, input: DateTime) -> Self

Only backups created after this time are listed. TimeRangeLowerBound is inclusive.

pub fn set_time_range_lower_bound(self, input: Option<DateTime>) -> Self

Only backups created after this time are listed. TimeRangeLowerBound is inclusive.

pub fn time_range_upper_bound(self, input: DateTime) -> Self

Only backups created before this time are listed. TimeRangeUpperBound is exclusive.

pub fn set_time_range_upper_bound(self, input: Option<DateTime>) -> Self

Only backups created before this time are listed. TimeRangeUpperBound is exclusive.

pub fn exclusive_start_backup_arn(self, input: impl Into<String>) -> Self

LastEvaluatedBackupArn is the Amazon Resource Name (ARN) of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

pub fn set_exclusive_start_backup_arn(self, input: Option<String>) -> Self

LastEvaluatedBackupArn is the Amazon Resource Name (ARN) of the backup last evaluated when the current page of results was returned, inclusive of the current page of results. This value may be specified as the ExclusiveStartBackupArn of a new ListBackups operation in order to fetch the next page of results.

pub fn backup_type(self, input: BackupTypeFilter) -> Self

The backups from the table specified by BackupType are listed.

Where BackupType can be:

USER - On-demand backup created by you. (The default setting if no other backup types are specified.)

SYSTEM - On-demand backup automatically created by DynamoDB.

ALL - All types of on-demand backups (USER and SYSTEM).

pub fn set_backup_type(self, input: Option<BackupTypeFilter>) -> Self

The backups from the table specified by BackupType are listed.

Where BackupType can be:

USER - On-demand backup created by you. (The default setting if no other backup types are specified.)

SYSTEM - On-demand backup automatically created by DynamoDB.

ALL - All types of on-demand backups (USER and SYSTEM).