Struct aws_sdk_dynamodb::client::fluent_builders::RestoreTableFromBackup

pub struct RestoreTableFromBackup<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> RestoreTableFromBackup<C>

pub async fn send(
self
) -> Result<RestoreTableFromBackupOutput, SdkError<RestoreTableFromBackupError>> where
C: SmithyConnector,

pub fn target_table_name(self, input: impl Into<String>) -> Self

The name of the new table to which the backup must be restored.

pub fn set_target_table_name(self, input: Option<String>) -> Self

pub fn backup_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) associated with the backup.

pub fn set_backup_arn(self, input: Option<String>) -> Self

pub fn billing_mode_override(self, input: BillingMode) -> Self

The billing mode of the restored table.

pub fn set_billing_mode_override(self, input: Option<BillingMode>) -> Self

pub fn global_secondary_index_override(
self,
inp: impl Into<GlobalSecondaryIndex>
) -> Self

List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

pub fn set_global_secondary_index_override(
self,
input: Option<Vec<GlobalSecondaryIndex>>
) -> Self

pub fn local_secondary_index_override(
self,
inp: impl Into<LocalSecondaryIndex>
) -> Self

List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

pub fn set_local_secondary_index_override(
self,
input: Option<Vec<LocalSecondaryIndex>>
) -> Self

pub fn provisioned_throughput_override(
self,
input: ProvisionedThroughput
) -> Self

Provisioned throughput settings for the restored table.

pub fn set_provisioned_throughput_override(
self,
input: Option<ProvisionedThroughput>
) -> Self

pub fn sse_specification_override(self, input: SseSpecification) -> Self

The new server-side encryption settings for the restored table.

pub fn set_sse_specification_override(
self,
input: Option<SseSpecification>
) -> Self