Struct aws_sdk_dynamodb::client::fluent_builders::UpdateTable

pub struct UpdateTable<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> UpdateTable<C>

pub async fn send(self) -> Result<UpdateTableOutput, SdkError<UpdateTableError>> where
C: SmithyConnector,

pub fn attribute_definitions(self, inp: impl Into<AttributeDefinition>) -> Self

An array of attributes that describe the key schema for the table and indexes. If you are adding a new global secondary index to the table, AttributeDefinitions must include the key element(s) of the new index.

pub fn set_attribute_definitions(
self,
input: Option<Vec<AttributeDefinition>>
) -> Self

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table to be updated.

pub fn set_table_name(self, input: Option<String>) -> Self

pub fn billing_mode(self, input: BillingMode) -> Self

Controls how you are charged for read and write throughput and how you manage capacity. When switching from pay-per-request to provisioned capacity, initial provisioned capacity values must be set. The initial provisioned capacity values are estimated based on the consumed read and write capacity of your table and global secondary indexes over the past 30 minutes.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

pub fn set_billing_mode(self, input: Option<BillingMode>) -> Self

pub fn provisioned_throughput(self, input: ProvisionedThroughput) -> Self

The new provisioned throughput settings for the specified table or index.

pub fn set_provisioned_throughput(
self,
input: Option<ProvisionedThroughput>
) -> Self

pub fn global_secondary_index_updates(
self,
inp: impl Into<GlobalSecondaryIndexUpdate>
) -> Self

An array of one or more global secondary indexes for the table. For each index in the array, you can request one action:

Create - add a new global secondary index to the table.

Update - modify the provisioned throughput settings of an existing global secondary index.

Delete - remove a global secondary index from the table.

You can create or delete only one global secondary index per UpdateTable operation.

For more information, see Managing Global Secondary Indexes in the Amazon DynamoDB Developer Guide.

pub fn set_global_secondary_index_updates(
self,
input: Option<Vec<GlobalSecondaryIndexUpdate>>
) -> Self

pub fn stream_specification(self, input: StreamSpecification) -> Self

Represents the DynamoDB Streams configuration for the table.

You receive a ResourceInUseException if you try to enable a stream on a table that already has a stream, or if you try to disable a stream on a table that doesn't have a stream.

pub fn set_stream_specification(
self,
input: Option<StreamSpecification>
) -> Self

pub fn sse_specification(self, input: SseSpecification) -> Self

The new server-side encryption settings for the specified table.

pub fn set_sse_specification(self, input: Option<SseSpecification>) -> Self

pub fn replica_updates(self, inp: impl Into<ReplicationGroupUpdate>) -> Self

A list of replica update actions (create, delete, or update) for the table.

This property only applies to Version 2019.11.21 of global tables.

pub fn set_replica_updates(
self,
input: Option<Vec<ReplicationGroupUpdate>>
) -> Self