Struct aws_sdk_dynamodb::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for DynamoDB_20120810.

This client allows ergonomic access to a DynamoDB_20120810-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_execute_statement(&self) -> BatchExecuteStatement<C, M, R>

pub fn batch_get_item(&self) -> BatchGetItem<C, M, R>

pub fn batch_write_item(&self) -> BatchWriteItem<C, M, R>

pub fn create_backup(&self) -> CreateBackup<C, M, R>

pub fn create_global_table(&self) -> CreateGlobalTable<C, M, R>

pub fn create_table(&self) -> CreateTable<C, M, R>

pub fn delete_backup(&self) -> DeleteBackup<C, M, R>

pub fn delete_item(&self) -> DeleteItem<C, M, R>

pub fn delete_table(&self) -> DeleteTable<C, M, R>

pub fn describe_backup(&self) -> DescribeBackup<C, M, R>

pub fn describe_continuous_backups(&self) -> DescribeContinuousBackups<C, M, R>

pub fn describe_contributor_insights(
&self
) -> DescribeContributorInsights<C, M, R>

pub fn describe_endpoints(&self) -> DescribeEndpoints<C, M, R>

pub fn describe_export(&self) -> DescribeExport<C, M, R>

pub fn describe_global_table(&self) -> DescribeGlobalTable<C, M, R>

pub fn describe_global_table_settings(
&self
) -> DescribeGlobalTableSettings<C, M, R>

pub fn describe_kinesis_streaming_destination(
&self
) -> DescribeKinesisStreamingDestination<C, M, R>

pub fn describe_limits(&self) -> DescribeLimits<C, M, R>

pub fn describe_table(&self) -> DescribeTable<C, M, R>

pub fn describe_table_replica_auto_scaling(
&self
) -> DescribeTableReplicaAutoScaling<C, M, R>

pub fn describe_time_to_live(&self) -> DescribeTimeToLive<C, M, R>

pub fn disable_kinesis_streaming_destination(
&self
) -> DisableKinesisStreamingDestination<C, M, R>

pub fn enable_kinesis_streaming_destination(
&self
) -> EnableKinesisStreamingDestination<C, M, R>

pub fn execute_statement(&self) -> ExecuteStatement<C, M, R>

pub fn execute_transaction(&self) -> ExecuteTransaction<C, M, R>

pub fn export_table_to_point_in_time(&self) -> ExportTableToPointInTime<C, M, R>

pub fn get_item(&self) -> GetItem<C, M, R>

pub fn list_backups(&self) -> ListBackups<C, M, R>

pub fn list_contributor_insights(&self) -> ListContributorInsights<C, M, R>

pub fn list_exports(&self) -> ListExports<C, M, R>

pub fn list_global_tables(&self) -> ListGlobalTables<C, M, R>

pub fn list_tables(&self) -> ListTables<C, M, R>

pub fn list_tags_of_resource(&self) -> ListTagsOfResource<C, M, R>

pub fn put_item(&self) -> PutItem<C, M, R>

pub fn query(&self) -> Query<C, M, R>

pub fn restore_table_from_backup(&self) -> RestoreTableFromBackup<C, M, R>

pub fn restore_table_to_point_in_time(
&self
) -> RestoreTableToPointInTime<C, M, R>

pub fn scan(&self) -> Scan<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn transact_get_items(&self) -> TransactGetItems<C, M, R>

pub fn transact_write_items(&self) -> TransactWriteItems<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_continuous_backups(&self) -> UpdateContinuousBackups<C, M, R>

pub fn update_contributor_insights(&self) -> UpdateContributorInsights<C, M, R>

pub fn update_global_table(&self) -> UpdateGlobalTable<C, M, R>

pub fn update_global_table_settings(&self) -> UpdateGlobalTableSettings<C, M, R>

pub fn update_item(&self) -> UpdateItem<C, M, R>

pub fn update_table(&self) -> UpdateTable<C, M, R>

pub fn update_table_replica_auto_scaling(
&self
) -> UpdateTableReplicaAutoScaling<C, M, R>

pub fn update_time_to_live(&self) -> UpdateTimeToLive<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self