Module aws_sdk_dynamodb::error

Modules

backup_in_use_exception

See BackupInUseException

backup_not_found_exception

See BackupNotFoundException

conditional_check_failed_exception

See ConditionalCheckFailedException

continuous_backups_unavailable_exception

See ContinuousBackupsUnavailableException

duplicate_item_exception

See DuplicateItemException

export_conflict_exception

See ExportConflictException

export_not_found_exception

See ExportNotFoundException

global_table_already_exists_exception

See GlobalTableAlreadyExistsException

global_table_not_found_exception

See GlobalTableNotFoundException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

index_not_found_exception

See IndexNotFoundException

internal_server_error

See InternalServerError

invalid_endpoint_exception

See InvalidEndpointException

invalid_export_time_exception

See InvalidExportTimeException

invalid_restore_time_exception

See InvalidRestoreTimeException

item_collection_size_limit_exceeded_exception

See ItemCollectionSizeLimitExceededException

limit_exceeded_exception

See LimitExceededException

point_in_time_recovery_unavailable_exception

See PointInTimeRecoveryUnavailableException

provisioned_throughput_exceeded_exception

See ProvisionedThroughputExceededException

replica_already_exists_exception

See ReplicaAlreadyExistsException

replica_not_found_exception

See ReplicaNotFoundException

request_limit_exceeded

See RequestLimitExceeded

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

table_already_exists_exception

See TableAlreadyExistsException

table_in_use_exception

See TableInUseException

table_not_found_exception

See TableNotFoundException

transaction_canceled_exception

See TransactionCanceledException

transaction_conflict_exception

See TransactionConflictException

transaction_in_progress_exception

See TransactionInProgressException

Structs

BackupInUseException

There is another ongoing conflicting backup control plane operation on the table. The backup is either being created, deleted or restored to a table.

BackupNotFoundException

Backup not found for the given BackupARN.

BatchExecuteStatementError

BatchGetItemError

BatchWriteItemError

ConditionalCheckFailedException

A condition specified in the operation could not be evaluated.

ContinuousBackupsUnavailableException

Backups have not yet been enabled for this table.

CreateBackupError

CreateGlobalTableError

CreateTableError

DeleteBackupError

DeleteItemError

DeleteTableError

DescribeBackupError

DescribeContinuousBackupsError

DescribeContributorInsightsError

DescribeEndpointsError

DescribeExportError

DescribeGlobalTableError

DescribeGlobalTableSettingsError

DescribeKinesisStreamingDestinationError

DescribeLimitsError

DescribeTableError

DescribeTableReplicaAutoScalingError

DescribeTimeToLiveError

DisableKinesisStreamingDestinationError

DuplicateItemException

EnableKinesisStreamingDestinationError

ExecuteStatementError

ExecuteTransactionError

ExportConflictException

There was a conflict when writing to the specified S3 bucket.

ExportNotFoundException

The specified export was not found.

ExportTableToPointInTimeError

GetItemError

GlobalTableAlreadyExistsException

The specified global table already exists.

GlobalTableNotFoundException

The specified global table does not exist.

IdempotentParameterMismatchException

DynamoDB rejected the request because you retried a request with a different payload but

IndexNotFoundException

The operation tried to access a nonexistent index.

InternalServerError

An error occurred on the server side.

InvalidEndpointException

InvalidExportTimeException

The specified ExportTime is outside of the point in time recovery

InvalidRestoreTimeException

An invalid restore time was specified. RestoreDateTime must be between EarliestRestorableDateTime and LatestRestorableDateTime.

ItemCollectionSizeLimitExceededException

An item collection is too large. This exception is only returned for tables that have one or more local secondary indexes.

LimitExceededException

There is no limit to the number of daily on-demand backups that can be taken.

ListBackupsError

ListContributorInsightsError

ListExportsError

ListGlobalTablesError

ListTablesError

ListTagsOfResourceError

PointInTimeRecoveryUnavailableException

Point in time recovery has not yet been enabled for this source table.

ProvisionedThroughputExceededException

Your request rate is too high. The AWS SDKs for DynamoDB automatically retry requests that

PutItemError

QueryError

ReplicaAlreadyExistsException

The specified replica is already part of the global table.

ReplicaNotFoundException

The specified replica is no longer part of the global table.

RequestLimitExceeded

Throughput exceeds the current throughput quota for your account. Please contact AWS Support at AWS Support to request a quota increase.

ResourceInUseException

The operation conflicts with the resource's availability. For example, you attempted to

ResourceNotFoundException

The operation tried to access a nonexistent table or index. The resource might not be specified

RestoreTableFromBackupError

RestoreTableToPointInTimeError

ScanError

TableAlreadyExistsException

A target table with the specified name already exists.

TableInUseException

A target table with the specified name is either being created or deleted.

TableNotFoundException

A source table with the name TableName does not currently exist within the subscriber's account.

TagResourceError

TransactGetItemsError

TransactWriteItemsError

TransactionCanceledException

The entire transaction request was canceled.

TransactionConflictException

Operation was rejected because there is an ongoing transaction for the item.

TransactionInProgressException

The transaction with the given request token is already in progress.

UntagResourceError

UpdateContinuousBackupsError

UpdateContributorInsightsError

UpdateGlobalTableError

UpdateGlobalTableSettingsError

UpdateItemError

UpdateTableError

UpdateTableReplicaAutoScalingError

UpdateTimeToLiveError

Enums

BatchExecuteStatementErrorKind

BatchGetItemErrorKind

BatchWriteItemErrorKind

CreateBackupErrorKind

CreateGlobalTableErrorKind

CreateTableErrorKind

DeleteBackupErrorKind

DeleteItemErrorKind

DeleteTableErrorKind

DescribeBackupErrorKind

DescribeContinuousBackupsErrorKind

DescribeContributorInsightsErrorKind

DescribeEndpointsErrorKind

DescribeExportErrorKind

DescribeGlobalTableErrorKind

DescribeGlobalTableSettingsErrorKind

DescribeKinesisStreamingDestinationErrorKind

DescribeLimitsErrorKind

DescribeTableErrorKind

DescribeTableReplicaAutoScalingErrorKind

DescribeTimeToLiveErrorKind

DisableKinesisStreamingDestinationErrorKind

EnableKinesisStreamingDestinationErrorKind

ExecuteStatementErrorKind

ExecuteTransactionErrorKind

ExportTableToPointInTimeErrorKind

GetItemErrorKind

ListBackupsErrorKind

ListContributorInsightsErrorKind

ListExportsErrorKind

ListGlobalTablesErrorKind

ListTablesErrorKind

ListTagsOfResourceErrorKind

PutItemErrorKind

QueryErrorKind

RestoreTableFromBackupErrorKind

RestoreTableToPointInTimeErrorKind

ScanErrorKind

TagResourceErrorKind

TransactGetItemsErrorKind

TransactWriteItemsErrorKind

UntagResourceErrorKind

UpdateContinuousBackupsErrorKind

UpdateContributorInsightsErrorKind

UpdateGlobalTableErrorKind

UpdateGlobalTableSettingsErrorKind

UpdateItemErrorKind

UpdateTableErrorKind

UpdateTableReplicaAutoScalingErrorKind

UpdateTimeToLiveErrorKind