Struct aws_sdk_dynamodb::input::export_table_to_point_in_time_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ExportTableToPointInTimeInput

Implementations

impl Builder

pub fn table_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) associated with the table to export.

pub fn set_table_arn(self, inp: Option<String>) -> Self

pub fn export_time(self, inp: Instant) -> Self

Time in the past from which to export table data. The table export will be a snapshot of the table's state at this point in time.

pub fn set_export_time(self, inp: Option<Instant>) -> Self

pub fn client_token(self, inp: impl Into<String>) -> Self

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an IdempotentParameterMismatch exception.

pub fn set_client_token(self, inp: Option<String>) -> Self

pub fn s3_bucket(self, inp: impl Into<String>) -> Self

The name of the Amazon S3 bucket to export the snapshot to.

pub fn set_s3_bucket(self, inp: Option<String>) -> Self

pub fn s3_bucket_owner(self, inp: impl Into<String>) -> Self

The ID of the AWS account that owns the bucket the export will be stored in.

pub fn set_s3_bucket_owner(self, inp: Option<String>) -> Self

pub fn s3_prefix(self, inp: impl Into<String>) -> Self

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

pub fn set_s3_prefix(self, inp: Option<String>) -> Self

pub fn s3_sse_algorithm(self, inp: S3SseAlgorithm) -> Self

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

pub fn set_s3_sse_algorithm(self, inp: Option<S3SseAlgorithm>) -> Self

pub fn s3_sse_kms_key_id(self, inp: impl Into<String>) -> Self

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

pub fn set_s3_sse_kms_key_id(self, inp: Option<String>) -> Self

pub fn export_format(self, inp: ExportFormat) -> Self

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn set_export_format(self, inp: Option<ExportFormat>) -> Self

pub fn build(self) -> Result<ExportTableToPointInTimeInput, BuildError>

Consumes the builder and constructs a ExportTableToPointInTimeInput