Struct aws_sdk_dynamodb::input::scan_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ScanInput

Implementations

impl Builder

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table containing the requested items; or, if you provide IndexName, the name of the table to which that index belongs.

pub fn set_table_name(self, input: Option<String>) -> Self

pub fn index_name(self, input: impl Into<String>) -> Self

The name of a secondary index to scan. This index can be any local secondary index or global secondary index. Note that if you use the IndexName parameter, you must also provide TableName.

pub fn set_index_name(self, input: Option<String>) -> Self

pub fn attributes_to_get(self, input: impl Into<String>) -> Self

pub fn set_attributes_to_get(self, input: Option<Vec<String>>) -> Self

pub fn limit(self, input: i32) -> Self

The maximum number of items to evaluate (not necessarily the number of matching items). If DynamoDB processes the number of items up to the limit while processing the results, it stops the operation and returns the matching values up to that point, and a key in LastEvaluatedKey to apply in a subsequent operation, so that you can pick up where you left off. Also, if the processed dataset size exceeds 1 MB before DynamoDB reaches this limit, it stops the operation and returns the matching values up to the limit, and a key in LastEvaluatedKey to apply in a subsequent operation to continue the operation. For more information, see Working with Queries in the Amazon DynamoDB Developer Guide.

pub fn set_limit(self, input: Option<i32>) -> Self

pub fn select(self, input: Select) -> Self

The attributes to be returned in the result. You can retrieve all item attributes, specific item attributes, the count of matching items, or in the case of an index, some or all of the attributes projected into the index.

ALL_ATTRIBUTES - Returns all of the item attributes from the specified table or index. If you query a local secondary index, then for each matching item in the index, DynamoDB fetches the entire item from the parent table. If the index is configured to project all item attributes, then all of the data can be obtained from the local secondary index, and no fetching is required.

ALL_PROJECTED_ATTRIBUTES - Allowed only when querying an index. Retrieves all attributes that have been projected into the index. If the index is configured to project all attributes, this return value is equivalent to specifying ALL_ATTRIBUTES.

COUNT - Returns the number of matching items, rather than the matching items themselves.

SPECIFIC_ATTRIBUTES - Returns only the attributes listed in AttributesToGet. This return value is equivalent to specifying AttributesToGet without specifying any value for Select.

If you query or scan a local secondary index and request only attributes that are projected into that index, the operation reads only the index and not the table. If any of the requested attributes are not projected into the local secondary index, DynamoDB fetches each of these attributes from the parent table. This extra fetching incurs additional throughput cost and latency.

If you query or scan a global secondary index, you can only request attributes that are projected into the index. Global secondary index queries cannot fetch attributes from the parent table.

If neither Select nor AttributesToGet are specified, DynamoDB defaults to ALL_ATTRIBUTES when accessing a table, and ALL_PROJECTED_ATTRIBUTES when accessing an index. You cannot use both Select and AttributesToGet together in a single request, unless the value for Select is SPECIFIC_ATTRIBUTES. (This usage is equivalent to specifying AttributesToGet without any value for Select.)

If you use the ProjectionExpression parameter, then the value for Select can only be SPECIFIC_ATTRIBUTES. Any other value for Select will return an error.

pub fn set_select(self, input: Option<Select>) -> Self

pub fn scan_filter(self, k: impl Into<String>, v: impl Into<Condition>) -> Self

pub fn set_scan_filter(self, input: Option<HashMap<String, Condition>>) -> Self

pub fn conditional_operator(self, input: ConditionalOperator) -> Self

This is a legacy parameter. Use FilterExpression instead. For more information, see ConditionalOperator in the Amazon DynamoDB Developer Guide.

pub fn set_conditional_operator(
self,
input: Option<ConditionalOperator>
) -> Self

pub fn exclusive_start_key(
self,
k: impl Into<String>,
v: impl Into<AttributeValue>
) -> Self

pub fn set_exclusive_start_key(
self,
input: Option<HashMap<String, AttributeValue>>
) -> Self

pub fn return_consumed_capacity(self, input: ReturnConsumedCapacity) -> Self

Determines the level of detail about provisioned throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn set_return_consumed_capacity(
self,
input: Option<ReturnConsumedCapacity>
) -> Self

pub fn total_segments(self, input: i32) -> Self

For a parallel Scan request, TotalSegments represents the total number of segments into which the Scan operation will be divided. The value of TotalSegments corresponds to the number of application workers that will perform the parallel scan. For example, if you want to use four application threads to scan a table or an index, specify a TotalSegments value of 4.

The value for TotalSegments must be greater than or equal to 1, and less than or equal to 1000000. If you specify a TotalSegments value of 1, the Scan operation will be sequential rather than parallel.

If you specify TotalSegments, you must also specify Segment.

pub fn set_total_segments(self, input: Option<i32>) -> Self

pub fn segment(self, input: i32) -> Self

For a parallel Scan request, Segment identifies an individual segment to be scanned by an application worker.

Segment IDs are zero-based, so the first segment is always 0. For example, if you want to use four application threads to scan a table or an index, then the first thread specifies a Segment value of 0, the second thread specifies 1, and so on.

The value of LastEvaluatedKey returned from a parallel Scan request must be used as ExclusiveStartKey with the same segment ID in a subsequent Scan operation.

The value for Segment must be greater than or equal to 0, and less than the value provided for TotalSegments.

If you provide Segment, you must also provide TotalSegments.

pub fn set_segment(self, input: Option<i32>) -> Self

pub fn projection_expression(self, input: impl Into<String>) -> Self

A string that identifies one or more attributes to retrieve from the specified table or index. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes will be returned. If any of the requested attributes are not found, they will not appear in the result.

For more information, see Specifying Item Attributes in the Amazon DynamoDB Developer Guide.

pub fn set_projection_expression(self, input: Option<String>) -> Self

pub fn filter_expression(self, input: impl Into<String>) -> Self

A string that contains conditions that DynamoDB applies after the Scan operation, but before the data is returned to you. Items that do not satisfy the FilterExpression criteria are not returned.

A FilterExpression is applied after the items have already been read; the process of filtering does not consume any additional read capacity units.

For more information, see Filter Expressions in the Amazon DynamoDB Developer Guide.

pub fn set_filter_expression(self, input: Option<String>) -> Self

pub fn expression_attribute_names(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

pub fn set_expression_attribute_names(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn expression_attribute_values(
self,
k: impl Into<String>,
v: impl Into<AttributeValue>
) -> Self

pub fn set_expression_attribute_values(
self,
input: Option<HashMap<String, AttributeValue>>
) -> Self

pub fn consistent_read(self, input: bool) -> Self

A Boolean value that determines the read consistency model during the scan:

If ConsistentRead is false, then the data returned from Scan might not contain the results from other recently completed write operations (PutItem, UpdateItem, or DeleteItem).

If ConsistentRead is true, then all of the write operations that completed before the Scan began are guaranteed to be contained in the Scan response.

The default setting for ConsistentRead is false.

The ConsistentRead parameter is not supported on global secondary indexes. If you scan a global secondary index with ConsistentRead set to true, you will receive a ValidationException.

pub fn set_consistent_read(self, input: Option<bool>) -> Self

pub fn build(self) -> Result<ScanInput, BuildError>

Consumes the builder and constructs a ScanInput