Struct aws_sdk_dynamodb::input::RestoreTableToPointInTimeInput

#[non_exhaustive]
pub struct RestoreTableToPointInTimeInput {
    pub source_table_arn: Option<String>,
    pub source_table_name: Option<String>,
    pub target_table_name: Option<String>,
    pub use_latest_restorable_time: Option<bool>,
    pub restore_date_time: Option<DateTime>,
    pub billing_mode_override: Option<BillingMode>,
    pub global_secondary_index_override: Option<Vec<GlobalSecondaryIndex>>,
    pub local_secondary_index_override: Option<Vec<LocalSecondaryIndex>>,
    pub provisioned_throughput_override: Option<ProvisionedThroughput>,
    pub sse_specification_override: Option<SseSpecification>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

source_table_arn: Option<String>

The DynamoDB table that will be restored. This value is an Amazon Resource Name (ARN).

source_table_name: Option<String>

Name of the source table that is being restored.

target_table_name: Option<String>

The name of the new table to which it must be restored to.

use_latest_restorable_time: Option<bool>

Restore the table to the latest possible time. LatestRestorableDateTime is typically 5 minutes before the current time.

restore_date_time: Option<DateTime>

Time in the past to restore the table to.

billing_mode_override: Option<BillingMode>

The billing mode of the restored table.

global_secondary_index_override: Option<Vec<GlobalSecondaryIndex>>

List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

local_secondary_index_override: Option<Vec<LocalSecondaryIndex>>

List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

provisioned_throughput_override: Option<ProvisionedThroughput>

Provisioned throughput settings for the restored table.

sse_specification_override: Option<SseSpecification>

The new server-side encryption settings for the restored table.

Implementations

impl RestoreTableToPointInTimeInput

pub async fn make_operation( &self, _config: &Config ) -> Result<Operation<RestoreTableToPointInTime, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<RestoreTableToPointInTime>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RestoreTableToPointInTimeInput

impl RestoreTableToPointInTimeInput

pub fn source_table_arn(&self) -> Option<&str>

The DynamoDB table that will be restored. This value is an Amazon Resource Name (ARN).

pub fn source_table_name(&self) -> Option<&str>

Name of the source table that is being restored.

pub fn target_table_name(&self) -> Option<&str>

The name of the new table to which it must be restored to.

pub fn use_latest_restorable_time(&self) -> Option<bool>

Restore the table to the latest possible time. LatestRestorableDateTime is typically 5 minutes before the current time.

pub fn restore_date_time(&self) -> Option<&DateTime>

Time in the past to restore the table to.

pub fn billing_mode_override(&self) -> Option<&BillingMode>

The billing mode of the restored table.

pub fn global_secondary_index_override(&self) -> Option<&[GlobalSecondaryIndex]>

List of global secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

pub fn local_secondary_index_override(&self) -> Option<&[LocalSecondaryIndex]>

List of local secondary indexes for the restored table. The indexes provided should match existing secondary indexes. You can choose to exclude some or all of the indexes at the time of restore.

pub fn provisioned_throughput_override(&self) -> Option<&ProvisionedThroughput>

Provisioned throughput settings for the restored table.

pub fn sse_specification_override(&self) -> Option<&SseSpecification>

The new server-side encryption settings for the restored table.