Struct aws_sdk_dynamodb::input::update_table_replica_auto_scaling_input::Builder

#[non_exhaustive]
pub struct Builder { /* private fields */ }

A builder for UpdateTableReplicaAutoScalingInput

Implementations

impl Builder

pub fn global_secondary_index_updates(
self,
input: GlobalSecondaryIndexAutoScalingUpdate
) -> Self

Appends an item to global_secondary_index_updates.

To override the contents of this collection use set_global_secondary_index_updates.

Represents the auto scaling settings of the global secondary indexes of the replica to be updated.

pub fn set_global_secondary_index_updates(
self,
input: Option<Vec<GlobalSecondaryIndexAutoScalingUpdate>>
) -> Self

Represents the auto scaling settings of the global secondary indexes of the replica to be updated.

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the global table to be updated.

pub fn set_table_name(self, input: Option<String>) -> Self

The name of the global table to be updated.

pub fn provisioned_write_capacity_auto_scaling_update(
self,
input: AutoScalingSettingsUpdate
) -> Self

Represents the auto scaling settings to be modified for a global table or global secondary index.

pub fn set_provisioned_write_capacity_auto_scaling_update(
self,
input: Option<AutoScalingSettingsUpdate>
) -> Self

Represents the auto scaling settings to be modified for a global table or global secondary index.

pub fn replica_updates(self, input: ReplicaAutoScalingUpdate) -> Self

Appends an item to replica_updates.

To override the contents of this collection use set_replica_updates.

Represents the auto scaling settings of replicas of the table that will be modified.

pub fn set_replica_updates(
self,
input: Option<Vec<ReplicaAutoScalingUpdate>>
) -> Self

Represents the auto scaling settings of replicas of the table that will be modified.

pub fn build(self) -> Result<UpdateTableReplicaAutoScalingInput, BuildError>

Consumes the builder and constructs a UpdateTableReplicaAutoScalingInput