Enum aws_sdk_dynamodb::model::AttributeValue

#[non_exhaustive]
pub enum AttributeValue {
    B(Blob),
    Bool(bool),
    Bs(Vec<Blob>),
    L(Vec<AttributeValue>),
    M(HashMap<String, AttributeValue>),
    N(String),
    Ns(Vec<String>),
    Null(bool),
    S(String),
    Ss(Vec<String>),
    Unknown,
}

Represents the data for an attribute.

Each attribute value is described as a name-value pair. The name is the data type, and the value is the data itself.

For more information, see Data Types in the Amazon DynamoDB Developer Guide.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

B(Blob)

An attribute of type Binary. For example:

"B": "dGhpcyB0ZXh0IGlzIGJhc2U2NC1lbmNvZGVk"

Bool(bool)

An attribute of type Boolean. For example:

"BOOL": true

Bs(Vec<Blob>)

An attribute of type Binary Set. For example:

"BS": ["U3Vubnk=", "UmFpbnk=", "U25vd3k="]

L(Vec<AttributeValue>)

An attribute of type List. For example:

"L": [ {"S": "Cookies"} , {"S": "Coffee"}, {"N": "3.14159"}]

M(HashMap<String, AttributeValue>)

An attribute of type Map. For example:

"M": {"Name": {"S": "Joe"}, "Age": {"N": "35"}}

N(String)

An attribute of type Number. For example:

"N": "123.45"

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

Ns(Vec<String>)

An attribute of type Number Set. For example:

"NS": ["42.2", "-19", "7.5", "3.14"]

Numbers are sent across the network to DynamoDB as strings, to maximize compatibility across languages and libraries. However, DynamoDB treats them as number type attributes for mathematical operations.

Null(bool)

An attribute of type Null. For example:

"NULL": true

S(String)

An attribute of type String. For example:

"S": "Hello"

Ss(Vec<String>)

An attribute of type String Set. For example:

"SS": ["Giraffe", "Hippo" ,"Zebra"]

Unknown

The Unknown variant represents cases where new union variant was received. Consider upgrading the SDK to the latest available version. An unknown enum variant

Note: If you encounter this error, consider upgrading your SDK to the latest version. The Unknown variant represents cases where the server sent a value that wasn’t recognized by the client. This can happen when the server adds new functionality, but the client has not been updated. To investigate this, consider turning on debug logging to print the raw HTTP response.

Implementations

impl AttributeValue

pub fn as_b(&self) -> Result<&Blob, &Self>

Tries to convert the enum instance into B, extracting the inner Blob. Returns Err(&Self) if it can’t be converted.

pub fn is_b(&self) -> bool

Returns true if this is a B.

pub fn as_bool(&self) -> Result<&bool, &Self>

Tries to convert the enum instance into Bool, extracting the inner bool. Returns Err(&Self) if it can’t be converted.

pub fn is_bool(&self) -> bool

Returns true if this is a Bool.

pub fn as_bs(&self) -> Result<&Vec<Blob>, &Self>

Tries to convert the enum instance into Bs, extracting the inner Vec. Returns Err(&Self) if it can’t be converted.

pub fn is_bs(&self) -> bool

Returns true if this is a Bs.

pub fn as_l(&self) -> Result<&Vec<AttributeValue>, &Self>

Tries to convert the enum instance into L, extracting the inner Vec. Returns Err(&Self) if it can’t be converted.

pub fn is_l(&self) -> bool

Returns true if this is a L.

pub fn as_m(&self) -> Result<&HashMap<String, AttributeValue>, &Self>

Tries to convert the enum instance into M, extracting the inner HashMap. Returns Err(&Self) if it can’t be converted.

pub fn is_m(&self) -> bool

Returns true if this is a M.

pub fn as_n(&self) -> Result<&String, &Self>

Tries to convert the enum instance into N, extracting the inner String. Returns Err(&Self) if it can’t be converted.

pub fn is_n(&self) -> bool

Returns true if this is a N.

pub fn as_ns(&self) -> Result<&Vec<String>, &Self>

Tries to convert the enum instance into Ns, extracting the inner Vec. Returns Err(&Self) if it can’t be converted.

pub fn is_ns(&self) -> bool

Returns true if this is a Ns.

pub fn as_null(&self) -> Result<&bool, &Self>

Tries to convert the enum instance into Null, extracting the inner bool. Returns Err(&Self) if it can’t be converted.

pub fn is_null(&self) -> bool

Returns true if this is a Null.

pub fn as_s(&self) -> Result<&String, &Self>

Tries to convert the enum instance into S, extracting the inner String. Returns Err(&Self) if it can’t be converted.

pub fn is_s(&self) -> bool

Returns true if this is a S.

pub fn as_ss(&self) -> Result<&Vec<String>, &Self>

Tries to convert the enum instance into Ss, extracting the inner Vec. Returns Err(&Self) if it can’t be converted.

pub fn is_ss(&self) -> bool

Returns true if this is a Ss.

pub fn is_unknown(&self) -> bool

Returns true if the enum instance is the Unknown variant.