Struct aws_sdk_dynamodb::model::export_description::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for ExportDescription

Implementations

impl Builder

pub fn export_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the table export.

pub fn set_export_arn(self, inp: Option<String>) -> Self

pub fn export_status(self, inp: ExportStatus) -> Self

Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

pub fn set_export_status(self, inp: Option<ExportStatus>) -> Self

pub fn start_time(self, inp: Instant) -> Self

The time at which the export task began.

pub fn set_start_time(self, inp: Option<Instant>) -> Self

pub fn end_time(self, inp: Instant) -> Self

The time at which the export task completed.

pub fn set_end_time(self, inp: Option<Instant>) -> Self

pub fn export_manifest(self, inp: impl Into<String>) -> Self

The name of the manifest file for the export task.

pub fn set_export_manifest(self, inp: Option<String>) -> Self

pub fn table_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the table that was exported.

pub fn set_table_arn(self, inp: Option<String>) -> Self

pub fn table_id(self, inp: impl Into<String>) -> Self

Unique ID of the table that was exported.

pub fn set_table_id(self, inp: Option<String>) -> Self

pub fn export_time(self, inp: Instant) -> Self

Point in time from which table data was exported.

pub fn set_export_time(self, inp: Option<Instant>) -> Self

pub fn client_token(self, inp: impl Into<String>) -> Self

The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

pub fn set_client_token(self, inp: Option<String>) -> Self

pub fn s3_bucket(self, inp: impl Into<String>) -> Self

The name of the Amazon S3 bucket containing the export.

pub fn set_s3_bucket(self, inp: Option<String>) -> Self

pub fn s3_bucket_owner(self, inp: impl Into<String>) -> Self

The ID of the AWS account that owns the bucket containing the export.

pub fn set_s3_bucket_owner(self, inp: Option<String>) -> Self

pub fn s3_prefix(self, inp: impl Into<String>) -> Self

The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

pub fn set_s3_prefix(self, inp: Option<String>) -> Self

pub fn s3_sse_algorithm(self, inp: S3SseAlgorithm) -> Self

Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with AWS KMS managed keys

pub fn set_s3_sse_algorithm(self, inp: Option<S3SseAlgorithm>) -> Self

pub fn s3_sse_kms_key_id(self, inp: impl Into<String>) -> Self

The ID of the AWS KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

pub fn set_s3_sse_kms_key_id(self, inp: Option<String>) -> Self

pub fn failure_code(self, inp: impl Into<String>) -> Self

Status code for the result of the failed export.

pub fn set_failure_code(self, inp: Option<String>) -> Self

pub fn failure_message(self, inp: impl Into<String>) -> Self

Export failure reason description.

pub fn set_failure_message(self, inp: Option<String>) -> Self

pub fn export_format(self, inp: ExportFormat) -> Self

The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn set_export_format(self, inp: Option<ExportFormat>) -> Self

pub fn billed_size_bytes(self, inp: i64) -> Self

The billable size of the table export.

pub fn set_billed_size_bytes(self, inp: Option<i64>) -> Self

pub fn item_count(self, inp: i64) -> Self

The number of items exported.

pub fn set_item_count(self, inp: Option<i64>) -> Self

pub fn build(self) -> ExportDescription

Consumes the builder and constructs a ExportDescription