Struct aws_sdk_dynamodb::model::Projection

#[non_exhaustive]
pub struct Projection {
    pub projection_type: Option<ProjectionType>,
    pub non_key_attributes: Option<Vec<String>>,
}

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

projection_type: Option<ProjectionType>

The set of attributes that are projected into the index:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - In addition to the attributes described in KEYS_ONLY, the secondary index will include other non-key attributes that you specify.

ALL - All of the table attributes are projected into the index.

non_key_attributes: Option<Vec<String>>

Represents the non-key attribute names which will be projected into the index.

For local secondary indexes, the total count of NonKeyAttributes summed across all of the local secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

Implementations

impl Projection

pub fn projection_type(&self) -> Option<&ProjectionType>

The set of attributes that are projected into the index:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - In addition to the attributes described in KEYS_ONLY, the secondary index will include other non-key attributes that you specify.

ALL - All of the table attributes are projected into the index.

pub fn non_key_attributes(&self) -> Option<&[String]>

Represents the non-key attribute names which will be projected into the index.

For local secondary indexes, the total count of NonKeyAttributes summed across all of the local secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

impl Projection

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Projection