Struct aws_sdk_dynamodb::model::ReplicaDescription

#[non_exhaustive]
pub struct ReplicaDescription {
    pub region_name: Option<String>,
    pub replica_status: Option<ReplicaStatus>,
    pub replica_status_description: Option<String>,
    pub replica_status_percent_progress: Option<String>,
    pub kms_master_key_id: Option<String>,
    pub provisioned_throughput_override: Option<ProvisionedThroughputOverride>,
    pub global_secondary_indexes: Option<Vec<ReplicaGlobalSecondaryIndexDescription>>,
    pub replica_inaccessible_date_time: Option<Instant>,
}

Contains the details of the replica.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

region_name: Option<String>

The name of the Region.

replica_status: Option<ReplicaStatus>

The current state of the replica:

CREATING - The replica is being created.

UPDATING - The replica is being updated.

DELETING - The replica is being deleted.

ACTIVE - The replica is ready for use.

REGION_DISABLED - The replica is inaccessible because the AWS Region has been disabled.

If the AWS Region remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The AWS KMS key used to encrypt the table is inaccessible.

If the AWS KMS key remains inaccessible for more than 20 hours, DynamoDB will remove this replica from the replication group. The replica will not be deleted and replication will stop from and to this region.

replica_status_description: Option<String>

Detailed information about the replica status.

replica_status_percent_progress: Option<String>

Specifies the progress of a Create, Update, or Delete action on the replica as a percentage.

kms_master_key_id: Option<String>

The AWS KMS customer master key (CMK) of the replica that will be used for AWS KMS encryption.

provisioned_throughput_override: Option<ProvisionedThroughputOverride>

Replica-specific provisioned throughput. If not described, uses the source table's provisioned throughput settings.

global_secondary_indexes: Option<Vec<ReplicaGlobalSecondaryIndexDescription>>

Replica-specific global secondary index settings.

replica_inaccessible_date_time: Option<Instant>

The time at which the replica was first detected as inaccessible. To determine cause of inaccessibility check the ReplicaStatus property.

Implementations

impl ReplicaDescription

pub fn builder() -> Builder

Creates a new builder-style object to manufacture ReplicaDescription