Struct aws_sdk_dynamodb::model::table_description::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for TableDescription

Implementations

impl Builder

pub fn attribute_definitions(
self,
input: impl Into<AttributeDefinition>
) -> Self

pub fn set_attribute_definitions(
self,
input: Option<Vec<AttributeDefinition>>
) -> Self

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table.

pub fn set_table_name(self, input: Option<String>) -> Self

pub fn key_schema(self, input: impl Into<KeySchemaElement>) -> Self

pub fn set_key_schema(self, input: Option<Vec<KeySchemaElement>>) -> Self

pub fn table_status(self, input: TableStatus) -> Self

The current state of the table:

CREATING - The table is being created.

UPDATING - The table is being updated.

DELETING - The table is being deleted.

ACTIVE - The table is ready for use.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The AWS KMS key used to encrypt the table in inaccessible. Table operations may fail due to failure to use the AWS KMS key. DynamoDB will initiate the table archival process when a table's AWS KMS key remains inaccessible for more than seven days.

ARCHIVING - The table is being archived. Operations are not allowed until archival is complete.

ARCHIVED - The table has been archived. See the ArchivalReason for more information.

pub fn set_table_status(self, input: Option<TableStatus>) -> Self

pub fn creation_date_time(self, input: Instant) -> Self

The date and time when the table was created, in UNIX epoch time format.

pub fn set_creation_date_time(self, input: Option<Instant>) -> Self

pub fn provisioned_throughput(
self,
input: ProvisionedThroughputDescription
) -> Self

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

pub fn set_provisioned_throughput(
self,
input: Option<ProvisionedThroughputDescription>
) -> Self

pub fn table_size_bytes(self, input: i64) -> Self

The total size of the specified table, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn set_table_size_bytes(self, input: Option<i64>) -> Self

pub fn item_count(self, input: i64) -> Self

The number of items in the specified table. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn set_item_count(self, input: Option<i64>) -> Self

pub fn table_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the table.

pub fn set_table_arn(self, input: Option<String>) -> Self

pub fn table_id(self, input: impl Into<String>) -> Self

Unique identifier for the table for which the backup was created.

pub fn set_table_id(self, input: Option<String>) -> Self

pub fn billing_mode_summary(self, input: BillingModeSummary) -> Self

Contains the details for the read/write capacity mode.

pub fn set_billing_mode_summary(self, input: Option<BillingModeSummary>) -> Self

pub fn local_secondary_indexes(
self,
input: impl Into<LocalSecondaryIndexDescription>
) -> Self

pub fn set_local_secondary_indexes(
self,
input: Option<Vec<LocalSecondaryIndexDescription>>
) -> Self

pub fn global_secondary_indexes(
self,
input: impl Into<GlobalSecondaryIndexDescription>
) -> Self

pub fn set_global_secondary_indexes(
self,
input: Option<Vec<GlobalSecondaryIndexDescription>>
) -> Self

pub fn stream_specification(self, input: StreamSpecification) -> Self

The current DynamoDB Streams configuration for the table.

pub fn set_stream_specification(
self,
input: Option<StreamSpecification>
) -> Self

pub fn latest_stream_label(self, input: impl Into<String>) -> Self

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

AWS customer ID

Table name

StreamLabel

pub fn set_latest_stream_label(self, input: Option<String>) -> Self

pub fn latest_stream_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the latest stream for this table.

pub fn set_latest_stream_arn(self, input: Option<String>) -> Self

pub fn global_table_version(self, input: impl Into<String>) -> Self

Represents the version of global tables in use, if the table is replicated across AWS Regions.

pub fn set_global_table_version(self, input: Option<String>) -> Self

pub fn replicas(self, input: impl Into<ReplicaDescription>) -> Self

pub fn set_replicas(self, input: Option<Vec<ReplicaDescription>>) -> Self

pub fn restore_summary(self, input: RestoreSummary) -> Self

Contains details for the restore.

pub fn set_restore_summary(self, input: Option<RestoreSummary>) -> Self

pub fn sse_description(self, input: SseDescription) -> Self

The description of the server-side encryption status on the specified table.

pub fn set_sse_description(self, input: Option<SseDescription>) -> Self

pub fn archival_summary(self, input: ArchivalSummary) -> Self

Contains information about the table archive.

pub fn set_archival_summary(self, input: Option<ArchivalSummary>) -> Self

pub fn build(self) -> TableDescription

Consumes the builder and constructs a TableDescription