Module aws_sdk_dynamodb::operation

Structs

BatchExecuteStatement

BatchGetItem

The BatchGetItem operation returns the attributes of one or more items from one or

BatchWriteItem

The BatchWriteItem operation puts or deletes multiple items in one or more

CreateBackup

Creates a backup for an existing table.

CreateGlobalTable

Creates a global table from an existing table. A global table creates a replication

CreateTable

The CreateTable operation adds a new table to your account. In an AWS

DeleteBackup

Deletes an existing backup of a table.

DeleteItem

Deletes a single item in a table by primary key. You can perform a conditional delete operation that deletes the item if it exists, or if it has an expected attribute value.

DeleteTable

The DeleteTable operation deletes a table and all of its items. After a

DescribeBackup

Describes an existing backup of a table.

DescribeContinuousBackups

Checks the status of continuous backups and point in time recovery on the specified table.

DescribeContributorInsights

Returns information about contributor insights, for a given table or global secondary index.

DescribeEndpoints

Returns the regional endpoint information.

DescribeExport

Describes an existing table export.

DescribeGlobalTable

Returns information about the specified global table.

DescribeGlobalTableSettings

Describes Region-specific settings for a global table.

DescribeKinesisStreamingDestination

Returns information about the status of Kinesis streaming.

DescribeLimits

Returns the current provisioned-capacity quotas for your AWS account in a Region, both

DescribeTable

Returns information about the table, including the current status of the table, when it was created, the primary key schema, and any indexes on the table.

DescribeTableReplicaAutoScaling

Describes auto scaling settings across replicas of the global table at once.

DescribeTimeToLive

Gives a description of the Time to Live (TTL) status on the specified table.

DisableKinesisStreamingDestination

Stops replication from the DynamoDB table to the Kinesis data stream. This is done

EnableKinesisStreamingDestination

Starts table data replication to the specified Kinesis data stream at a timestamp chosen

ExecuteStatement

ExecuteTransaction

ExportTableToPointInTime

Exports table data to an S3 bucket. The table must have point in time recovery

GetItem

The GetItem operation returns a set of attributes for the item with the given primary

ListBackups

List backups associated with an AWS account. To list backups for a given table, specify

ListContributorInsights

Returns a list of ContributorInsightsSummary for a table and all its global secondary indexes.

ListExports

Lists completed exports within the past 90 days.

ListGlobalTables

Lists all global tables that have a replica in the specified Region.

ListTables

Returns an array of table names associated with the current account and endpoint. The output

ListTagsOfResource

List all tags on an Amazon DynamoDB resource. You can call ListTagsOfResource up to 10 times per second, per account.

PutItem

Creates a new item, or replaces an old item with a new item. If an item that has the same primary key as the new item already exists in the specified table, the new item completely replaces the existing item. You can perform a conditional put operation (add a new item if one with the specified primary key doesn't exist), or replace an existing item if it has certain attribute values. You can return the item's attribute values in the same operation, using the ReturnValues parameter.

Query

The Query operation finds items based on primary key values.

RestoreTableFromBackup

Creates a new table from an existing backup. Any number of users can execute up to 4 concurrent restores

RestoreTableToPointInTime

Restores the specified table to the specified point in time within

Scan

The Scan operation returns one or more items and item attributes by accessing every

TagResource

Associate a set of tags with an Amazon DynamoDB resource. You can then activate these

TransactGetItems

TransactWriteItems

UntagResource

Removes the association of tags from an Amazon DynamoDB resource. You can call

UpdateContinuousBackups

UpdateContributorInsights

Updates the status for contributor insights for a specific table or index.

UpdateGlobalTable

Adds or removes replicas in the specified global table. The global table must already

UpdateGlobalTableSettings

Updates settings for a global table.

UpdateItem

Edits an existing item's attributes, or adds a new item to the table if it does not already exist. You can put, delete, or add attribute values. You can also perform a conditional update on an existing item (insert a new attribute name-value pair if it doesn't exist, or replace an existing name-value pair if it has certain expected attribute values).

UpdateTable

Modifies the provisioned throughput settings, global secondary indexes, or DynamoDB Streams settings for a given table.

UpdateTableReplicaAutoScaling

Updates auto scaling settings on your global tables at once.

UpdateTimeToLive

The UpdateTimeToLive method enables or disables Time to Live (TTL) for the