Struct aws_sdk_dynamodb::operation::batch_get_item::builders::BatchGetItemFluentBuilder

pub struct BatchGetItemFluentBuilder { /* private fields */ }

Fluent builder constructing a request to BatchGetItem.

The BatchGetItem operation returns the attributes of one or more items from one or more tables. You identify requested items by primary key.

A single operation can retrieve up to 16 MB of data, which can contain as many as 100 items. BatchGetItem returns a partial result if the response size limit is exceeded, the table's provisioned throughput is exceeded, more than 1MB per partition is requested, or an internal processing failure occurs. If a partial result is returned, the operation returns a value for UnprocessedKeys. You can use this value to retry the operation starting with the next item to get.

If you request more than 100 items, BatchGetItem returns a ValidationException with the message "Too many items requested for the BatchGetItem call."

For example, if you ask to retrieve 100 items, but each individual item is 300 KB in size, the system returns 52 items (so as not to exceed the 16 MB limit). It also returns an appropriate UnprocessedKeys value so you can get the next page of results. If desired, your application can include its own logic to assemble the pages of results into one dataset.

If none of the items can be processed due to insufficient provisioned throughput on all of the tables in the request, then BatchGetItem returns a ProvisionedThroughputExceededException. If at least one of the items is successfully processed, then BatchGetItem completes successfully, while returning the keys of the unread items in UnprocessedKeys.

If DynamoDB returns any unprocessed items, you should retry the batch operation on those items. However, we strongly recommend that you use an exponential backoff algorithm. If you retry the batch operation immediately, the underlying read or write requests can still fail due to throttling on the individual tables. If you delay the batch operation using exponential backoff, the individual requests in the batch are much more likely to succeed.

For more information, see Batch Operations and Error Handling in the Amazon DynamoDB Developer Guide.

By default, BatchGetItem performs eventually consistent reads on every table in the request. If you want strongly consistent reads instead, you can set ConsistentRead to true for any or all tables.

In order to minimize response latency, BatchGetItem may retrieve items in parallel.

When designing your application, keep in mind that DynamoDB does not return items in any particular order. To help parse the response by item, include the primary key values for the items in your request in the ProjectionExpression parameter.

If a requested item does not exist, it is not returned in the result. Requests for nonexistent items consume the minimum read capacity units according to the type of read. For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

Implementations

impl BatchGetItemFluentBuilder

pub fn as_input(&self) -> &BatchGetItemInputBuilder

Access the BatchGetItem as a reference.

pub async fn send( self, ) -> Result<BatchGetItemOutput, SdkError<BatchGetItemError, HttpResponse>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn customize( self, ) -> CustomizableOperation<BatchGetItemOutput, BatchGetItemError, Self>

Consumes this builder, creating a customizable operation that can be modified before being sent.

pub fn request_items(self, k: impl Into<String>, v: KeysAndAttributes) -> Self

Adds a key-value pair to RequestItems.

To override the contents of this collection use set_request_items.

A map of one or more table names or table ARNs and, for each table, a map that describes one or more items to retrieve from that table. Each table name or ARN can be used only once per BatchGetItem request.

Each element in the map of items to retrieve consists of the following:

ConsistentRead - If true, a strongly consistent read is used; if false (the default), an eventually consistent read is used.

ExpressionAttributeNames - One or more substitution tokens for attribute names in the ProjectionExpression parameter. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information about expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

Keys - An array of primary key attribute values that define specific items in the table. For each primary key, you must provide all of the key attributes. For example, with a simple primary key, you only need to provide the partition key value. For a composite key, you must provide both the partition key value and the sort key value.

ProjectionExpression - A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes are returned. If any of the requested attributes are not found, they do not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

AttributesToGet - This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide.

pub fn set_request_items( self, input: Option<HashMap<String, KeysAndAttributes>>, ) -> Self

A map of one or more table names or table ARNs and, for each table, a map that describes one or more items to retrieve from that table. Each table name or ARN can be used only once per BatchGetItem request.

Each element in the map of items to retrieve consists of the following:

ConsistentRead - If true, a strongly consistent read is used; if false (the default), an eventually consistent read is used.

ExpressionAttributeNames - One or more substitution tokens for attribute names in the ProjectionExpression parameter. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information about expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

Keys - An array of primary key attribute values that define specific items in the table. For each primary key, you must provide all of the key attributes. For example, with a simple primary key, you only need to provide the partition key value. For a composite key, you must provide both the partition key value and the sort key value.

ProjectionExpression - A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes are returned. If any of the requested attributes are not found, they do not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

AttributesToGet - This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide.

pub fn get_request_items(&self) -> &Option<HashMap<String, KeysAndAttributes>>

A map of one or more table names or table ARNs and, for each table, a map that describes one or more items to retrieve from that table. Each table name or ARN can be used only once per BatchGetItem request.

Each element in the map of items to retrieve consists of the following:

ConsistentRead - If true, a strongly consistent read is used; if false (the default), an eventually consistent read is used.

ExpressionAttributeNames - One or more substitution tokens for attribute names in the ProjectionExpression parameter. The following are some use cases for using ExpressionAttributeNames:

To access an attribute whose name conflicts with a DynamoDB reserved word.

To create a placeholder for repeating occurrences of an attribute name in an expression.

To prevent special characters in an attribute name from being misinterpreted in an expression.

Use the # character in an expression to dereference an attribute name. For example, consider the following attribute name:

Percentile

The name of this attribute conflicts with a reserved word, so it cannot be used directly in an expression. (For the complete list of reserved words, see Reserved Words in the Amazon DynamoDB Developer Guide). To work around this, you could specify the following for ExpressionAttributeNames:

{"#P":"Percentile"}

You could then use this substitution in an expression, as in this example:

#P = :val

Tokens that begin with the : character are expression attribute values, which are placeholders for the actual value at runtime.

For more information about expression attribute names, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

Keys - An array of primary key attribute values that define specific items in the table. For each primary key, you must provide all of the key attributes. For example, with a simple primary key, you only need to provide the partition key value. For a composite key, you must provide both the partition key value and the sort key value.

ProjectionExpression - A string that identifies one or more attributes to retrieve from the table. These attributes can include scalars, sets, or elements of a JSON document. The attributes in the expression must be separated by commas.

If no attribute names are specified, then all attributes are returned. If any of the requested attributes are not found, they do not appear in the result.

For more information, see Accessing Item Attributes in the Amazon DynamoDB Developer Guide.

AttributesToGet - This is a legacy parameter. Use ProjectionExpression instead. For more information, see AttributesToGet in the Amazon DynamoDB Developer Guide.

pub fn return_consumed_capacity(self, input: ReturnConsumedCapacity) -> Self

Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn set_return_consumed_capacity( self, input: Option<ReturnConsumedCapacity>, ) -> Self

Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.

pub fn get_return_consumed_capacity(&self) -> &Option<ReturnConsumedCapacity>

Determines the level of detail about either provisioned or on-demand throughput consumption that is returned in the response:

INDEXES - The response includes the aggregate ConsumedCapacity for the operation, together with ConsumedCapacity for each table and secondary index that was accessed.

Note that some operations, such as GetItem and BatchGetItem, do not access any indexes at all. In these cases, specifying INDEXES will only return ConsumedCapacity information for table(s).

TOTAL - The response includes only the aggregate ConsumedCapacity for the operation.

NONE - No ConsumedCapacity details are included in the response.