Struct aws_sdk_dynamodb::operation::create_table::builders::CreateTableFluentBuilder

pub struct CreateTableFluentBuilder { /* private fields */ }

Fluent builder constructing a request to CreateTable.

The CreateTable operation adds a new table to your account. In an Amazon Web Services account, table names must be unique within each Region. That is, you can have two tables with same name if you create the tables in different Regions.

CreateTable is an asynchronous operation. Upon receiving a CreateTable request, DynamoDB immediately returns a response with a TableStatus of CREATING. After the table is created, DynamoDB sets the TableStatus to ACTIVE. You can perform read and write operations only on an ACTIVE table.

You can optionally define secondary indexes on the new table, as part of the CreateTable operation. If you want to create multiple tables with secondary indexes on them, you must create the tables sequentially. Only one table with secondary indexes can be in the CREATING state at any given time.

You can use the DescribeTable action to check the table status.

Implementations

impl CreateTableFluentBuilder

pub fn as_input(&self) -> &CreateTableInputBuilder

Access the CreateTable as a reference.

pub async fn send( self, ) -> Result<CreateTableOutput, SdkError<CreateTableError, HttpResponse>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn customize( self, ) -> CustomizableOperation<CreateTableOutput, CreateTableError, Self>

Consumes this builder, creating a customizable operation that can be modified before being sent.

pub fn attribute_definitions(self, input: AttributeDefinition) -> Self

Appends an item to AttributeDefinitions.

To override the contents of this collection use set_attribute_definitions.

An array of attributes that describe the key schema for the table and indexes.

pub fn set_attribute_definitions( self, input: Option<Vec<AttributeDefinition>>, ) -> Self

An array of attributes that describe the key schema for the table and indexes.

pub fn get_attribute_definitions(&self) -> &Option<Vec<AttributeDefinition>>

An array of attributes that describe the key schema for the table and indexes.

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table to create. You can also provide the Amazon Resource Name (ARN) of the table in this parameter.

pub fn set_table_name(self, input: Option<String>) -> Self

The name of the table to create. You can also provide the Amazon Resource Name (ARN) of the table in this parameter.

pub fn get_table_name(&self) -> &Option<String>

The name of the table to create. You can also provide the Amazon Resource Name (ARN) of the table in this parameter.

pub fn key_schema(self, input: KeySchemaElement) -> Self

Appends an item to KeySchema.

To override the contents of this collection use set_key_schema.

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

pub fn set_key_schema(self, input: Option<Vec<KeySchemaElement>>) -> Self

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

pub fn get_key_schema(&self) -> &Option<Vec<KeySchemaElement>>

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

pub fn local_secondary_indexes(self, input: LocalSecondaryIndex) -> Self

Appends an item to LocalSecondaryIndexes.

To override the contents of this collection use set_local_secondary_indexes.

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

pub fn set_local_secondary_indexes( self, input: Option<Vec<LocalSecondaryIndex>>, ) -> Self

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

pub fn get_local_secondary_indexes(&self) -> &Option<Vec<LocalSecondaryIndex>>

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

pub fn global_secondary_indexes(self, input: GlobalSecondaryIndex) -> Self

Appends an item to GlobalSecondaryIndexes.

To override the contents of this collection use set_global_secondary_indexes.

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

pub fn set_global_secondary_indexes( self, input: Option<Vec<GlobalSecondaryIndex>>, ) -> Self

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

pub fn get_global_secondary_indexes(&self) -> &Option<Vec<GlobalSecondaryIndex>>

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

pub fn billing_mode(self, input: BillingMode) -> Self

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned capacity mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-demand capacity mode.

pub fn set_billing_mode(self, input: Option<BillingMode>) -> Self

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned capacity mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-demand capacity mode.

pub fn get_billing_mode(&self) -> &Option<BillingMode>

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned capacity mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-demand capacity mode.

pub fn provisioned_throughput(self, input: ProvisionedThroughput) -> Self

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

pub fn set_provisioned_throughput( self, input: Option<ProvisionedThroughput>, ) -> Self

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

pub fn get_provisioned_throughput(&self) -> &Option<ProvisionedThroughput>

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

pub fn stream_specification(self, input: StreamSpecification) -> Self

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

pub fn set_stream_specification( self, input: Option<StreamSpecification>, ) -> Self

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

pub fn get_stream_specification(&self) -> &Option<StreamSpecification>

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

pub fn sse_specification(self, input: SseSpecification) -> Self

Represents the settings used to enable server-side encryption.

pub fn set_sse_specification(self, input: Option<SseSpecification>) -> Self

Represents the settings used to enable server-side encryption.

pub fn get_sse_specification(&self) -> &Option<SseSpecification>

Represents the settings used to enable server-side encryption.

pub fn tags(self, input: Tag) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

pub fn get_tags(&self) -> &Option<Vec<Tag>>

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

pub fn table_class(self, input: TableClass) -> Self

The table class of the new table. Valid values are STANDARD and STANDARD_INFREQUENT_ACCESS.

pub fn set_table_class(self, input: Option<TableClass>) -> Self

The table class of the new table. Valid values are STANDARD and STANDARD_INFREQUENT_ACCESS.

pub fn get_table_class(&self) -> &Option<TableClass>

The table class of the new table. Valid values are STANDARD and STANDARD_INFREQUENT_ACCESS.

pub fn deletion_protection_enabled(self, input: bool) -> Self

Indicates whether deletion protection is to be enabled (true) or disabled (false) on the table.

pub fn set_deletion_protection_enabled(self, input: Option<bool>) -> Self

Indicates whether deletion protection is to be enabled (true) or disabled (false) on the table.

pub fn get_deletion_protection_enabled(&self) -> &Option<bool>

Indicates whether deletion protection is to be enabled (true) or disabled (false) on the table.

pub fn resource_policy(self, input: impl Into<String>) -> Self

An Amazon Web Services resource-based policy document in JSON format that will be attached to the table.

When you attach a resource-based policy while creating a table, the policy application is strongly consistent.

The maximum size supported for a resource-based policy document is 20 KB. DynamoDB counts whitespaces when calculating the size of a policy against this limit. For a full list of all considerations that apply for resource-based policies, see Resource-based policy considerations.

You need to specify the CreateTable and PutResourcePolicy IAM actions for authorizing a user to create a table with a resource-based policy.

pub fn set_resource_policy(self, input: Option<String>) -> Self

An Amazon Web Services resource-based policy document in JSON format that will be attached to the table.

When you attach a resource-based policy while creating a table, the policy application is strongly consistent.

The maximum size supported for a resource-based policy document is 20 KB. DynamoDB counts whitespaces when calculating the size of a policy against this limit. For a full list of all considerations that apply for resource-based policies, see Resource-based policy considerations.

You need to specify the CreateTable and PutResourcePolicy IAM actions for authorizing a user to create a table with a resource-based policy.

pub fn get_resource_policy(&self) -> &Option<String>

An Amazon Web Services resource-based policy document in JSON format that will be attached to the table.

When you attach a resource-based policy while creating a table, the policy application is strongly consistent.

The maximum size supported for a resource-based policy document is 20 KB. DynamoDB counts whitespaces when calculating the size of a policy against this limit. For a full list of all considerations that apply for resource-based policies, see Resource-based policy considerations.

You need to specify the CreateTable and PutResourcePolicy IAM actions for authorizing a user to create a table with a resource-based policy.

pub fn on_demand_throughput(self, input: OnDemandThroughput) -> Self

Sets the maximum number of read and write units for the specified table in on-demand capacity mode. If you use this parameter, you must specify MaxReadRequestUnits, MaxWriteRequestUnits, or both.

pub fn set_on_demand_throughput(self, input: Option<OnDemandThroughput>) -> Self

Sets the maximum number of read and write units for the specified table in on-demand capacity mode. If you use this parameter, you must specify MaxReadRequestUnits, MaxWriteRequestUnits, or both.

pub fn get_on_demand_throughput(&self) -> &Option<OnDemandThroughput>

Sets the maximum number of read and write units for the specified table in on-demand capacity mode. If you use this parameter, you must specify MaxReadRequestUnits, MaxWriteRequestUnits, or both.