Struct aws_sdk_dynamodb::operation::export_table_to_point_in_time::builders::ExportTableToPointInTimeFluentBuilder

pub struct ExportTableToPointInTimeFluentBuilder { /* private fields */ }

Fluent builder constructing a request to ExportTableToPointInTime.

Exports table data to an S3 bucket. The table must have point in time recovery enabled, and you can export data from any time within the point in time recovery window.

Implementations

impl ExportTableToPointInTimeFluentBuilder

pub fn as_input(&self) -> &ExportTableToPointInTimeInputBuilder

Access the ExportTableToPointInTime as a reference.

pub async fn send( self, ) -> Result<ExportTableToPointInTimeOutput, SdkError<ExportTableToPointInTimeError, HttpResponse>>

Sends the request and returns the response.

If an error occurs, an SdkError will be returned with additional details that can be matched against.

By default, any retryable failures will be retried twice. Retry behavior is configurable with the RetryConfig, which can be set when configuring the client.

pub fn customize( self, ) -> CustomizableOperation<ExportTableToPointInTimeOutput, ExportTableToPointInTimeError, Self>

Consumes this builder, creating a customizable operation that can be modified before being sent.

pub fn table_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) associated with the table to export.

pub fn set_table_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) associated with the table to export.

pub fn get_table_arn(&self) -> &Option<String>

The Amazon Resource Name (ARN) associated with the table to export.

pub fn export_time(self, input: DateTime) -> Self

Time in the past from which to export table data, counted in seconds from the start of the Unix epoch. The table export will be a snapshot of the table's state at this point in time.

pub fn set_export_time(self, input: Option<DateTime>) -> Self

Time in the past from which to export table data, counted in seconds from the start of the Unix epoch. The table export will be a snapshot of the table's state at this point in time.

pub fn get_export_time(&self) -> &Option<DateTime>

Time in the past from which to export table data, counted in seconds from the start of the Unix epoch. The table export will be a snapshot of the table's state at this point in time.

pub fn client_token(self, input: impl Into<String>) -> Self

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an ImportConflictException.

pub fn set_client_token(self, input: Option<String>) -> Self

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an ImportConflictException.

pub fn get_client_token(&self) -> &Option<String>

Providing a ClientToken makes the call to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

A client token is valid for 8 hours after the first request that uses it is completed. After 8 hours, any request with the same client token is treated as a new request. Do not resubmit the same request with the same client token for more than 8 hours, or the result might not be idempotent.

If you submit a request with the same client token but a change in other parameters within the 8-hour idempotency window, DynamoDB returns an ImportConflictException.

pub fn s3_bucket(self, input: impl Into<String>) -> Self

The name of the Amazon S3 bucket to export the snapshot to.

pub fn set_s3_bucket(self, input: Option<String>) -> Self

The name of the Amazon S3 bucket to export the snapshot to.

pub fn get_s3_bucket(&self) -> &Option<String>

The name of the Amazon S3 bucket to export the snapshot to.

pub fn s3_bucket_owner(self, input: impl Into<String>) -> Self

The ID of the Amazon Web Services account that owns the bucket the export will be stored in.

S3BucketOwner is a required parameter when exporting to a S3 bucket in another account.

pub fn set_s3_bucket_owner(self, input: Option<String>) -> Self

The ID of the Amazon Web Services account that owns the bucket the export will be stored in.

S3BucketOwner is a required parameter when exporting to a S3 bucket in another account.

pub fn get_s3_bucket_owner(&self) -> &Option<String>

The ID of the Amazon Web Services account that owns the bucket the export will be stored in.

S3BucketOwner is a required parameter when exporting to a S3 bucket in another account.

pub fn s3_prefix(self, input: impl Into<String>) -> Self

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

pub fn set_s3_prefix(self, input: Option<String>) -> Self

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

pub fn get_s3_prefix(&self) -> &Option<String>

The Amazon S3 bucket prefix to use as the file name and path of the exported snapshot.

pub fn s3_sse_algorithm(self, input: S3SseAlgorithm) -> Self

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with KMS managed keys

pub fn set_s3_sse_algorithm(self, input: Option<S3SseAlgorithm>) -> Self

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with KMS managed keys

pub fn get_s3_sse_algorithm(&self) -> &Option<S3SseAlgorithm>

Type of encryption used on the bucket where export data will be stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with KMS managed keys

pub fn s3_sse_kms_key_id(self, input: impl Into<String>) -> Self

The ID of the KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

pub fn set_s3_sse_kms_key_id(self, input: Option<String>) -> Self

The ID of the KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

pub fn get_s3_sse_kms_key_id(&self) -> &Option<String>

The ID of the KMS managed key used to encrypt the S3 bucket where export data will be stored (if applicable).

pub fn export_format(self, input: ExportFormat) -> Self

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn set_export_format(self, input: Option<ExportFormat>) -> Self

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn get_export_format(&self) -> &Option<ExportFormat>

The format for the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn export_type(self, input: ExportType) -> Self

Choice of whether to execute as a full export or incremental export. Valid values are FULL_EXPORT or INCREMENTAL_EXPORT. The default value is FULL_EXPORT. If INCREMENTAL_EXPORT is provided, the IncrementalExportSpecification must also be used.

pub fn set_export_type(self, input: Option<ExportType>) -> Self

Choice of whether to execute as a full export or incremental export. Valid values are FULL_EXPORT or INCREMENTAL_EXPORT. The default value is FULL_EXPORT. If INCREMENTAL_EXPORT is provided, the IncrementalExportSpecification must also be used.

pub fn get_export_type(&self) -> &Option<ExportType>

Choice of whether to execute as a full export or incremental export. Valid values are FULL_EXPORT or INCREMENTAL_EXPORT. The default value is FULL_EXPORT. If INCREMENTAL_EXPORT is provided, the IncrementalExportSpecification must also be used.

pub fn incremental_export_specification( self, input: IncrementalExportSpecification, ) -> Self

Optional object containing the parameters specific to an incremental export.

pub fn set_incremental_export_specification( self, input: Option<IncrementalExportSpecification>, ) -> Self

Optional object containing the parameters specific to an incremental export.

pub fn get_incremental_export_specification( &self, ) -> &Option<IncrementalExportSpecification>

Optional object containing the parameters specific to an incremental export.