Enum aws_sdk_dynamodb::operation::export_table_to_point_in_time::ExportTableToPointInTimeError

#[non_exhaustive]
pub enum ExportTableToPointInTimeError {
    ExportConflictException(ExportConflictException),
    InternalServerError(InternalServerError),
    InvalidExportTimeException(InvalidExportTimeException),
    LimitExceededException(LimitExceededException),
    PointInTimeRecoveryUnavailableException(PointInTimeRecoveryUnavailableException),
    TableNotFoundException(TableNotFoundException),
    Unhandled(Unhandled),
}

Error type for the ExportTableToPointInTimeError operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ExportConflictException(ExportConflictException)

There was a conflict when writing to the specified S3 bucket.

InternalServerError(InternalServerError)

An error occurred on the server side.

InvalidExportTimeException(InvalidExportTimeException)

The specified ExportTime is outside of the point in time recovery window.

LimitExceededException(LimitExceededException)

There is no limit to the number of daily on-demand backups that can be taken.

For most purposes, up to 500 simultaneous table operations are allowed per account. These operations include CreateTable, UpdateTable, DeleteTable,UpdateTimeToLive, RestoreTableFromBackup, and RestoreTableToPointInTime.

When you are creating a table with one or more secondary indexes, you can have up to 250 such requests running at a time. However, if the table or index specifications are complex, then DynamoDB might temporarily reduce the number of concurrent operations.

When importing into DynamoDB, up to 50 simultaneous import table operations are allowed per account.

There is a soft account quota of 2,500 tables.

GetRecords was called with a value of more than 1000 for the limit request parameter.

More than 2 processes are reading from the same streams shard at the same time. Exceeding this limit may result in request throttling.

PointInTimeRecoveryUnavailableException(PointInTimeRecoveryUnavailableException)

Point in time recovery has not yet been enabled for this source table.

TableNotFoundException(TableNotFoundException)

A source table with the name TableName does not currently exist within the subscriber's account or the subscriber is operating in the wrong Amazon Web Services Region.

Unhandled(Unhandled)

๐Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): ยยยerr if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).

Implementations

impl ExportTableToPointInTimeError

pub fn unhandled(err: impl Into<Box<dyn Error + Send + Sync + 'static>>) -> Self

Creates the ExportTableToPointInTimeError::Unhandled variant from any error type.

pub fn generic(err: ErrorMetadata) -> Self

Creates the ExportTableToPointInTimeError::Unhandled variant from an ErrorMetadata.

pub fn meta(&self) -> &ErrorMetadata

Returns error metadata, which includes the error code, message, request ID, and potentially additional information.

pub fn is_export_conflict_exception(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::ExportConflictException.

pub fn is_internal_server_error(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::InternalServerError.

pub fn is_invalid_export_time_exception(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::InvalidExportTimeException.

pub fn is_limit_exceeded_exception(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::LimitExceededException.

pub fn is_point_in_time_recovery_unavailable_exception(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::PointInTimeRecoveryUnavailableException.

pub fn is_table_not_found_exception(&self) -> bool

Returns true if the error kind is ExportTableToPointInTimeError::TableNotFoundException.