aws_sdk_dynamodb::operation::list_backups

Enum ListBackupsError

#[non_exhaustive]
pub enum ListBackupsError {
    InternalServerError(InternalServerError),
    InvalidEndpointException(InvalidEndpointException),
    Unhandled(Unhandled),
}

Error type for the ListBackupsError operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

InternalServerError(InternalServerError)

An error occurred on the server side.

InvalidEndpointException(InvalidEndpointException)

Unhandled(Unhandled)

๐Deprecated: Matching Unhandled directly is not forwards compatible. Instead, match using a variable wildcard pattern and check .code(): ยยยerr if err.code() == Some("SpecificExceptionCode") => { /* handle the error */ } See ProvideErrorMetadata for what information is available for the error.

An unexpected error occurred (e.g., invalid JSON returned by the service or an unknown error code).

Implementations

impl ListBackupsError

pub fn unhandled(err: impl Into<Box<dyn Error + Send + Sync + 'static>>) -> Self

Creates the ListBackupsError::Unhandled variant from any error type.

pub fn generic(err: ErrorMetadata) -> Self

Creates the ListBackupsError::Unhandled variant from an ErrorMetadata.

pub fn meta(&self) -> &ErrorMetadata

Returns error metadata, which includes the error code, message, request ID, and potentially additional information.

pub fn is_internal_server_error(&self) -> bool

Returns true if the error kind is ListBackupsError::InternalServerError.

pub fn is_invalid_endpoint_exception(&self) -> bool

Returns true if the error kind is ListBackupsError::InvalidEndpointException.