Struct aws_sdk_dynamodb::paginator::QueryPaginator

pub struct QueryPaginator { /* private fields */ }

Paginator for Query

Implementations

impl QueryPaginator

pub fn page_size(self, limit: i32) -> Self

Set the page size

Note: this method will override any previously set value for limit

pub fn items(self) -> QueryPaginatorItems

Create a flattened paginator

This paginator automatically flattens results using items. Queries to the underlying service are dispatched lazily.

pub fn send( self ) -> impl Stream<Item = Result<QueryOutput, SdkError<QueryError>>> + Unpin

Create the pagination stream

Note: No requests will be dispatched until the stream is used (eg. with .next().await).