Module aws_sdk_dynamodb::types::error

Error types that Amazon DynamoDB can respond with.

Modules

builders

Builders

Structs

BackupInUseException

There is another ongoing conflicting backup control plane operation on the table. The backup is either being created, deleted or restored to a table.

BackupNotFoundException

Backup not found for the given BackupARN.

ConditionalCheckFailedException

A condition specified in the operation could not be evaluated.

ContinuousBackupsUnavailableException

Backups have not yet been enabled for this table.

DuplicateItemException

There was an attempt to insert an item with the same primary key as an item that already exists in the DynamoDB table.

ExportConflictException

There was a conflict when writing to the specified S3 bucket.

ExportNotFoundException

The specified export was not found.

GlobalTableAlreadyExistsException

The specified global table already exists.

GlobalTableNotFoundException

The specified global table does not exist.

IdempotentParameterMismatchException

DynamoDB rejected the request because you retried a request with a different payload but with an idempotent token that was already used.

ImportConflictException

There was a conflict when importing from the specified S3 source. This can occur when the current import conflicts with a previous import request that had the same client token.

ImportNotFoundException

The specified import was not found.

IndexNotFoundException

The operation tried to access a nonexistent index.

InternalServerError

An error occurred on the server side.

InvalidEndpointException

InvalidExportTimeException

The specified ExportTime is outside of the point in time recovery window.

InvalidRestoreTimeException

An invalid restore time was specified. RestoreDateTime must be between EarliestRestorableDateTime and LatestRestorableDateTime.

ItemCollectionSizeLimitExceededException

An item collection is too large. This exception is only returned for tables that have one or more local secondary indexes.

LimitExceededException

There is no limit to the number of daily on-demand backups that can be taken.

PointInTimeRecoveryUnavailableException

Point in time recovery has not yet been enabled for this source table.

ProvisionedThroughputExceededException

Your request rate is too high. The Amazon Web Services SDKs for DynamoDB automatically retry requests that receive this exception. Your request is eventually successful, unless your retry queue is too large to finish. Reduce the frequency of requests and use exponential backoff. For more information, go to Error Retries and Exponential Backoff in the Amazon DynamoDB Developer Guide.

ReplicaAlreadyExistsException

The specified replica is already part of the global table.

ReplicaNotFoundException

The specified replica is no longer part of the global table.

RequestLimitExceeded

Throughput exceeds the current throughput quota for your account. Please contact Amazon Web Services Support to request a quota increase.

ResourceInUseException

The operation conflicts with the resource's availability. For example, you attempted to recreate an existing table, or tried to delete a table currently in the CREATING state.

ResourceNotFoundException

The operation tried to access a nonexistent table or index. The resource might not be specified correctly, or its status might not be ACTIVE.

TableAlreadyExistsException

A target table with the specified name already exists.

TableInUseException

A target table with the specified name is either being created or deleted.

TableNotFoundException

A source table with the name TableName does not currently exist within the subscriber's account or the subscriber is operating in the wrong Amazon Web Services Region.

TransactionCanceledException

The entire transaction request was canceled.

TransactionConflictException

Operation was rejected because there is an ongoing transaction for the item.

TransactionInProgressException

The transaction with the given request token is already in progress.