Struct aws_sdk_dynamodb::types::ExportDescription

#[non_exhaustive]
pub struct ExportDescription {

    pub export_arn: Option<String>,
    pub export_status: Option<ExportStatus>,
    pub start_time: Option<DateTime>,
    pub end_time: Option<DateTime>,
    pub export_manifest: Option<String>,
    pub table_arn: Option<String>,
    pub table_id: Option<String>,
    pub export_time: Option<DateTime>,
    pub client_token: Option<String>,
    pub s3_bucket: Option<String>,
    pub s3_bucket_owner: Option<String>,
    pub s3_prefix: Option<String>,
    pub s3_sse_algorithm: Option<S3SseAlgorithm>,
    pub s3_sse_kms_key_id: Option<String>,
    pub failure_code: Option<String>,
    pub failure_message: Option<String>,
    pub export_format: Option<ExportFormat>,
    pub billed_size_bytes: Option<i64>,
    pub item_count: Option<i64>,
    pub export_type: Option<ExportType>,
    pub incremental_export_specification: Option<IncrementalExportSpecification>,

}

Represents the properties of the exported table.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

export_arn: Option<String>

The Amazon Resource Name (ARN) of the table export.

export_status: Option<ExportStatus>

Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

start_time: Option<DateTime>

The time at which the export task began.

end_time: Option<DateTime>

The time at which the export task completed.

export_manifest: Option<String>

The name of the manifest file for the export task.

table_arn: Option<String>

The Amazon Resource Name (ARN) of the table that was exported.

table_id: Option<String>

Unique ID of the table that was exported.

export_time: Option<DateTime>

Point in time from which table data was exported.

client_token: Option<String>

The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

s3_bucket: Option<String>

The name of the Amazon S3 bucket containing the export.

s3_bucket_owner: Option<String>

The ID of the Amazon Web Services account that owns the bucket containing the export.

s3_prefix: Option<String>

The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

s3_sse_algorithm: Option<S3SseAlgorithm>

Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with KMS managed keys

s3_sse_kms_key_id: Option<String>

The ID of the KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

failure_code: Option<String>

Status code for the result of the failed export.

failure_message: Option<String>

Export failure reason description.

export_format: Option<ExportFormat>

The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

billed_size_bytes: Option<i64>

The billable size of the table export.

item_count: Option<i64>

The number of items exported.

export_type: Option<ExportType>

The type of export that was performed. Valid values are FULL_EXPORT or INCREMENTAL_EXPORT.

incremental_export_specification: Option<IncrementalExportSpecification>

Optional object containing the parameters specific to an incremental export.

Implementations

impl ExportDescription

pub fn export_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the table export.

pub fn export_status(&self) -> Option<&ExportStatus>

Export can be in one of the following states: IN_PROGRESS, COMPLETED, or FAILED.

pub fn start_time(&self) -> Option<&DateTime>

The time at which the export task began.

pub fn end_time(&self) -> Option<&DateTime>

The time at which the export task completed.

pub fn export_manifest(&self) -> Option<&str>

The name of the manifest file for the export task.

pub fn table_arn(&self) -> Option<&str>

The Amazon Resource Name (ARN) of the table that was exported.

pub fn table_id(&self) -> Option<&str>

Unique ID of the table that was exported.

pub fn export_time(&self) -> Option<&DateTime>

Point in time from which table data was exported.

pub fn client_token(&self) -> Option<&str>

The client token that was provided for the export task. A client token makes calls to ExportTableToPointInTimeInput idempotent, meaning that multiple identical calls have the same effect as one single call.

pub fn s3_bucket(&self) -> Option<&str>

The name of the Amazon S3 bucket containing the export.

pub fn s3_bucket_owner(&self) -> Option<&str>

The ID of the Amazon Web Services account that owns the bucket containing the export.

pub fn s3_prefix(&self) -> Option<&str>

The Amazon S3 bucket prefix used as the file name and path of the exported snapshot.

pub fn s3_sse_algorithm(&self) -> Option<&S3SseAlgorithm>

Type of encryption used on the bucket where export data is stored. Valid values for S3SseAlgorithm are:

AES256 - server-side encryption with Amazon S3 managed keys

KMS - server-side encryption with KMS managed keys

pub fn s3_sse_kms_key_id(&self) -> Option<&str>

The ID of the KMS managed key used to encrypt the S3 bucket where export data is stored (if applicable).

pub fn failure_code(&self) -> Option<&str>

Status code for the result of the failed export.

pub fn failure_message(&self) -> Option<&str>

Export failure reason description.

pub fn export_format(&self) -> Option<&ExportFormat>

The format of the exported data. Valid values for ExportFormat are DYNAMODB_JSON or ION.

pub fn billed_size_bytes(&self) -> Option<i64>

The billable size of the table export.

pub fn item_count(&self) -> Option<i64>

The number of items exported.

pub fn export_type(&self) -> Option<&ExportType>

The type of export that was performed. Valid values are FULL_EXPORT or INCREMENTAL_EXPORT.

pub fn incremental_export_specification( &self, ) -> Option<&IncrementalExportSpecification>

Optional object containing the parameters specific to an incremental export.

impl ExportDescription

pub fn builder() -> ExportDescriptionBuilder

Creates a new builder-style object to manufacture ExportDescription.