Struct aws_sdk_dynamodb::types::ReplicaGlobalSecondaryIndexDescription

#[non_exhaustive]
pub struct ReplicaGlobalSecondaryIndexDescription {
    pub index_name: Option<String>,
    pub provisioned_throughput_override: Option<ProvisionedThroughputOverride>,
    pub on_demand_throughput_override: Option<OnDemandThroughputOverride>,
}

Represents the properties of a replica global secondary index.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

index_name: Option<String>

The name of the global secondary index.

provisioned_throughput_override: Option<ProvisionedThroughputOverride>

If not described, uses the source table GSI's read capacity settings.

on_demand_throughput_override: Option<OnDemandThroughputOverride>

Overrides the maximum on-demand throughput for the specified global secondary index in the specified replica table.

Implementations

impl ReplicaGlobalSecondaryIndexDescription

pub fn index_name(&self) -> Option<&str>

The name of the global secondary index.

pub fn provisioned_throughput_override( &self, ) -> Option<&ProvisionedThroughputOverride>

If not described, uses the source table GSI's read capacity settings.

pub fn on_demand_throughput_override( &self, ) -> Option<&OnDemandThroughputOverride>

Overrides the maximum on-demand throughput for the specified global secondary index in the specified replica table.

impl ReplicaGlobalSecondaryIndexDescription

pub fn builder() -> ReplicaGlobalSecondaryIndexDescriptionBuilder

Creates a new builder-style object to manufacture ReplicaGlobalSecondaryIndexDescription.