Type Alias aws_sdk_dynamodb::waiters::table_not_exists::WaitUntilTableNotExistsError

pub type WaitUntilTableNotExistsError = WaiterError<DescribeTableOutput, DescribeTableError>;

Error type for the table_not_exists waiter.

Aliased Type

enum WaitUntilTableNotExistsError {
    ConstructionFailure(ConstructionFailure),
    ExceededMaxWait(ExceededMaxWait),
    FailureState(FailureState<DescribeTableOutput, DescribeTableError>),
    OperationFailed(OperationFailed<DescribeTableError>),
}

Variants

ConstructionFailure(ConstructionFailure)

An error occurred during waiter initialization.

This can happen if the input/config is invalid.

ExceededMaxWait(ExceededMaxWait)

The maximum wait time was exceeded without completion.

FailureState(FailureState<DescribeTableOutput, DescribeTableError>)

Waiting ended in a failure state.

A failed waiter state can occur on a successful response from the server if, for example, that response indicates that the thing being waited for won’t succeed/finish.

A failure state error will only occur for successful or modeled error responses. Unmodeled error responses will never make it into this error case.

OperationFailed(OperationFailed<DescribeTableError>)

A polling operation failed while waiting.

This error will only occur for unmodeled errors. Modeled errors can potentially be handled by the waiter logic, and will therefore end up in [WaiterError::FailureState].

Note: If retry is configured, this means that the operation failed after retrying the configured number of attempts.