Crate aws_sdk_dynamodb

Amazon DynamoDB

Amazon DynamoDB is a fully managed NoSQL database service that provides fast and predictable performance with seamless scalability. DynamoDB lets you offload the administrative burdens of operating and scaling a distributed database, so that you don't have to worry about hardware provisioning, setup and configuration, replication, software patching, or cluster scaling.

With DynamoDB, you can create database tables that can store and retrieve any amount of data, and serve any level of request traffic. You can scale up or scale down your tables' throughput capacity without downtime or performance degradation, and use the Amazon Web Services Management Console to monitor resource utilization and performance metrics.

DynamoDB automatically spreads the data and traffic for your tables over a sufficient number of servers to handle your throughput and storage requirements, while maintaining consistent and fast performance. All of your data is stored on solid state disks (SSDs) and automatically replicated across multiple Availability Zones in an Amazon Web Services Region, providing built-in high availability and data durability.

Crate Organization

The entry point for most customers will be Client. Client exposes one method for each API offered by the service.

Some APIs require complex or nested arguments. These exist in model.

Lastly, errors that can be returned by the service are contained within error. Error defines a meta error encompassing all possible errors that can be returned by the service.

The other modules within this crate are not required for normal usage.

Examples

Examples can be found here.

Modules

client

Client and fluent builders for calling the service.

config

Configuration for the service.

error

Errors that can occur when calling the service.

input

Input structures for operations.

lens

Generated accessors for nested fields

middleware

Base Middleware Stack

model

Data structures used by operation inputs/outputs.

operation

All operations that this crate can perform.

output

Output structures for operations.

paginator

Paginators for the service

types

Re-exported types from supporting crates.

Structs

AppName

App name that can be configured with an AWS SDK client to become part of the user agent string.

Client

Client for Amazon DynamoDB

Config

Service config.

Credentials

AWS SDK Credentials

Endpoint

API Endpoint

Region

The region to send requests to.

RetryConfig

Retry configuration for requests.

Enums

Error

All possible error types for this service.

Statics

PKG_VERSION

Crate version number.