Module aws_sdk_dynamodb::error

Modules

backup_in_use_error
See BackupInUseError

backup_not_found_error
See BackupNotFoundError

conditional_check_failed_error
See ConditionalCheckFailedError

continuous_backups_unavailable_error
See ContinuousBackupsUnavailableError

duplicate_item_error
See DuplicateItemError

export_conflict_error
See ExportConflictError

export_not_found_error
See ExportNotFoundError

global_table_already_exists_error
See GlobalTableAlreadyExistsError

global_table_not_found_error
See GlobalTableNotFoundError

idempotent_parameter_mismatch_error
See IdempotentParameterMismatchError

index_not_found_error
See IndexNotFoundError

internal_server_error
See InternalServerError

invalid_endpoint_error
See InvalidEndpointError

invalid_export_time_error
See InvalidExportTimeError

invalid_restore_time_error
See InvalidRestoreTimeError

item_collection_size_limit_exceeded_error
See ItemCollectionSizeLimitExceededError

limit_exceeded_error
See LimitExceededError

point_in_time_recovery_unavailable_error
See PointInTimeRecoveryUnavailableError

provisioned_throughput_exceeded_error
See ProvisionedThroughputExceededError

replica_already_exists_error
See ReplicaAlreadyExistsError

replica_not_found_error
See ReplicaNotFoundError

request_limit_exceeded
See RequestLimitExceeded

resource_in_use_error
See ResourceInUseError

resource_not_found_error
See ResourceNotFoundError

table_already_exists_error
See TableAlreadyExistsError

table_in_use_error
See TableInUseError

table_not_found_error
See TableNotFoundError

transaction_canceled_error
See TransactionCanceledError

transaction_conflict_error
See TransactionConflictError

transaction_in_progress_error
See TransactionInProgressError

Structs

BackupInUseError
There is another ongoing conflicting backup control plane operation on the table. The backup is either being created, deleted or restored to a table.

BackupNotFoundError
Backup not found for the given BackupARN.

BatchExecuteStatementError

BatchGetItemError

BatchWriteItemError

ConditionalCheckFailedError
A condition specified in the operation could not be evaluated.

ContinuousBackupsUnavailableError
Backups have not yet been enabled for this table.

CreateBackupError

CreateGlobalTableError

CreateTableError

DeleteBackupError

DeleteItemError

DeleteTableError

DescribeBackupError

DescribeContinuousBackupsError

DescribeContributorInsightsError

DescribeEndpointsError

DescribeExportError

DescribeGlobalTableError

DescribeGlobalTableSettingsError

DescribeKinesisStreamingDestinationError

DescribeLimitsError

DescribeTableError

DescribeTableReplicaAutoScalingError

DescribeTimeToLiveError

DisableKinesisStreamingDestinationError

DuplicateItemError

EnableKinesisStreamingDestinationError

ExecuteStatementError

ExecuteTransactionError

ExportConflictError
There was a conflict when writing to the specified S3 bucket.

ExportNotFoundError
The specified export was not found.

ExportTableToPointInTimeError

GetItemError

GlobalTableAlreadyExistsError
The specified global table already exists.

GlobalTableNotFoundError
The specified global table does not exist.

IdempotentParameterMismatchError
DynamoDB rejected the request because you retried a request with a different payload but

IndexNotFoundError
The operation tried to access a nonexistent index.

InternalServerError
An error occurred on the server side.

InvalidEndpointError

InvalidExportTimeError
The specified ExportTime is outside of the point in time recovery

InvalidRestoreTimeError
An invalid restore time was specified. RestoreDateTime must be between EarliestRestorableDateTime and LatestRestorableDateTime.

ItemCollectionSizeLimitExceededError
An item collection is too large. This exception is only returned for tables that have one or more local secondary indexes.

LimitExceededError
There is no limit to the number of daily on-demand backups that can be taken.

ListBackupsError

ListContributorInsightsError

ListExportsError

ListGlobalTablesError

ListTablesError

ListTagsOfResourceError

PointInTimeRecoveryUnavailableError
Point in time recovery has not yet been enabled for this source table.

ProvisionedThroughputExceededError
Your request rate is too high. The AWS SDKs for DynamoDB automatically retry requests that

PutItemError

QueryError

ReplicaAlreadyExistsError
The specified replica is already part of the global table.

ReplicaNotFoundError
The specified replica is no longer part of the global table.

RequestLimitExceeded
Throughput exceeds the current throughput quota for your account. Please contact AWS Support at AWS Support to request a quota increase.

ResourceInUseError
The operation conflicts with the resource's availability. For example, you attempted to

ResourceNotFoundError
The operation tried to access a nonexistent table or index. The resource might not be specified

RestoreTableFromBackupError

RestoreTableToPointInTimeError

ScanError

TableAlreadyExistsError
A target table with the specified name already exists.

TableInUseError
A target table with the specified name is either being created or deleted.

TableNotFoundError
A source table with the name TableName does not currently exist within the subscriber's account.

TagResourceError

TransactGetItemsError

TransactWriteItemsError

TransactionCanceledError
The entire transaction request was canceled.

TransactionConflictError
Operation was rejected because there is an ongoing transaction for the item.

TransactionInProgressError
The transaction with the given request token is already in progress.

UntagResourceError

UpdateContinuousBackupsError

UpdateContributorInsightsError

UpdateGlobalTableError

UpdateGlobalTableSettingsError

UpdateItemError

UpdateTableError

UpdateTableReplicaAutoScalingError

UpdateTimeToLiveError

Enums

BatchExecuteStatementErrorKind

BatchGetItemErrorKind

BatchWriteItemErrorKind

CreateBackupErrorKind

CreateGlobalTableErrorKind

CreateTableErrorKind

DeleteBackupErrorKind

DeleteItemErrorKind

DeleteTableErrorKind

DescribeBackupErrorKind

DescribeContinuousBackupsErrorKind

DescribeContributorInsightsErrorKind

DescribeEndpointsErrorKind

DescribeExportErrorKind

DescribeGlobalTableErrorKind

DescribeGlobalTableSettingsErrorKind

DescribeKinesisStreamingDestinationErrorKind

DescribeLimitsErrorKind

DescribeTableErrorKind

DescribeTableReplicaAutoScalingErrorKind

DescribeTimeToLiveErrorKind

DisableKinesisStreamingDestinationErrorKind

EnableKinesisStreamingDestinationErrorKind

ExecuteStatementErrorKind

ExecuteTransactionErrorKind

ExportTableToPointInTimeErrorKind

GetItemErrorKind

ListBackupsErrorKind

ListContributorInsightsErrorKind

ListExportsErrorKind

ListGlobalTablesErrorKind

ListTablesErrorKind

ListTagsOfResourceErrorKind

PutItemErrorKind

QueryErrorKind

RestoreTableFromBackupErrorKind

RestoreTableToPointInTimeErrorKind

ScanErrorKind

TagResourceErrorKind

TransactGetItemsErrorKind

TransactWriteItemsErrorKind

UntagResourceErrorKind

UpdateContinuousBackupsErrorKind

UpdateContributorInsightsErrorKind

UpdateGlobalTableErrorKind

UpdateGlobalTableSettingsErrorKind

UpdateItemErrorKind

UpdateTableErrorKind

UpdateTableReplicaAutoScalingErrorKind

UpdateTimeToLiveErrorKind