Module aws_sdk_dynamodb::error

Errors that can occur when calling the service.

Modules

backup_in_use_exception

See BackupInUseException

backup_not_found_exception

See BackupNotFoundException

conditional_check_failed_exception

See ConditionalCheckFailedException

continuous_backups_unavailable_exception

See ContinuousBackupsUnavailableException

duplicate_item_exception

See DuplicateItemException

export_conflict_exception

See ExportConflictException

export_not_found_exception

See ExportNotFoundException

global_table_already_exists_exception

See GlobalTableAlreadyExistsException

global_table_not_found_exception

See GlobalTableNotFoundException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

index_not_found_exception

See IndexNotFoundException

internal_server_error

See InternalServerError

invalid_endpoint_exception

See InvalidEndpointException

invalid_export_time_exception

See InvalidExportTimeException

invalid_restore_time_exception

See InvalidRestoreTimeException

item_collection_size_limit_exceeded_exception

See ItemCollectionSizeLimitExceededException

limit_exceeded_exception

See LimitExceededException

point_in_time_recovery_unavailable_exception

See PointInTimeRecoveryUnavailableException

provisioned_throughput_exceeded_exception

See ProvisionedThroughputExceededException

replica_already_exists_exception

See ReplicaAlreadyExistsException

replica_not_found_exception

See ReplicaNotFoundException

request_limit_exceeded

See RequestLimitExceeded

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

table_already_exists_exception

See TableAlreadyExistsException

table_in_use_exception

See TableInUseException

table_not_found_exception

See TableNotFoundException

transaction_canceled_exception

See TransactionCanceledException

transaction_conflict_exception

See TransactionConflictException

transaction_in_progress_exception

See TransactionInProgressException

Structs

BackupInUseException

There is another ongoing conflicting backup control plane operation on the table. The backup is either being created, deleted or restored to a table.

BackupNotFoundException

Backup not found for the given BackupARN.

BatchExecuteStatementError

Error type for the BatchExecuteStatement operation.

BatchGetItemError

Error type for the BatchGetItem operation.

BatchWriteItemError

Error type for the BatchWriteItem operation.

ConditionalCheckFailedException

A condition specified in the operation could not be evaluated.

ContinuousBackupsUnavailableException

Backups have not yet been enabled for this table.

CreateBackupError

Error type for the CreateBackup operation.

CreateGlobalTableError

Error type for the CreateGlobalTable operation.

CreateTableError

Error type for the CreateTable operation.

DeleteBackupError

Error type for the DeleteBackup operation.

DeleteItemError

Error type for the DeleteItem operation.

DeleteTableError

Error type for the DeleteTable operation.

DescribeBackupError

Error type for the DescribeBackup operation.

DescribeContinuousBackupsError

Error type for the DescribeContinuousBackups operation.

DescribeContributorInsightsError

Error type for the DescribeContributorInsights operation.

DescribeEndpointsError

Error type for the DescribeEndpoints operation.

DescribeExportError

Error type for the DescribeExport operation.

DescribeGlobalTableError

Error type for the DescribeGlobalTable operation.

DescribeGlobalTableSettingsError

Error type for the DescribeGlobalTableSettings operation.

DescribeKinesisStreamingDestinationError

Error type for the DescribeKinesisStreamingDestination operation.

DescribeLimitsError

Error type for the DescribeLimits operation.

DescribeTableError

Error type for the DescribeTable operation.

DescribeTableReplicaAutoScalingError

Error type for the DescribeTableReplicaAutoScaling operation.

DescribeTimeToLiveError

Error type for the DescribeTimeToLive operation.

DisableKinesisStreamingDestinationError

Error type for the DisableKinesisStreamingDestination operation.

DuplicateItemException

There was an attempt to insert an item with the same primary key as an item that already exists in the DynamoDB table.

EnableKinesisStreamingDestinationError

Error type for the EnableKinesisStreamingDestination operation.

ExecuteStatementError

Error type for the ExecuteStatement operation.

ExecuteTransactionError

Error type for the ExecuteTransaction operation.

ExportConflictException

There was a conflict when writing to the specified S3 bucket.

ExportNotFoundException

The specified export was not found.

ExportTableToPointInTimeError

Error type for the ExportTableToPointInTime operation.

GetItemError

Error type for the GetItem operation.

GlobalTableAlreadyExistsException

The specified global table already exists.

GlobalTableNotFoundException

The specified global table does not exist.

IdempotentParameterMismatchException

DynamoDB rejected the request because you retried a request with a different payload but with an idempotent token that was already used.

IndexNotFoundException

The operation tried to access a nonexistent index.

InternalServerError

An error occurred on the server side.

InvalidEndpointException

InvalidExportTimeException

The specified ExportTime is outside of the point in time recovery window.

InvalidRestoreTimeException

An invalid restore time was specified. RestoreDateTime must be between EarliestRestorableDateTime and LatestRestorableDateTime.

ItemCollectionSizeLimitExceededException

An item collection is too large. This exception is only returned for tables that have one or more local secondary indexes.

LimitExceededException

There is no limit to the number of daily on-demand backups that can be taken.

ListBackupsError

Error type for the ListBackups operation.

ListContributorInsightsError

Error type for the ListContributorInsights operation.

ListExportsError

Error type for the ListExports operation.

ListGlobalTablesError

Error type for the ListGlobalTables operation.

ListTablesError

Error type for the ListTables operation.

ListTagsOfResourceError

Error type for the ListTagsOfResource operation.

PointInTimeRecoveryUnavailableException

Point in time recovery has not yet been enabled for this source table.

ProvisionedThroughputExceededException

Your request rate is too high. The Amazon Web Services SDKs for DynamoDB automatically retry requests that receive this exception. Your request is eventually successful, unless your retry queue is too large to finish. Reduce the frequency of requests and use exponential backoff. For more information, go to Error Retries and Exponential Backoff in the Amazon DynamoDB Developer Guide.

PutItemError

Error type for the PutItem operation.

QueryError

Error type for the Query operation.

ReplicaAlreadyExistsException

The specified replica is already part of the global table.

ReplicaNotFoundException

The specified replica is no longer part of the global table.

RequestLimitExceeded

Throughput exceeds the current throughput quota for your account. Please contact Amazon Web Services Support to request a quota increase.

ResourceInUseException

The operation conflicts with the resource's availability. For example, you attempted to recreate an existing table, or tried to delete a table currently in the CREATING state.

ResourceNotFoundException

The operation tried to access a nonexistent table or index. The resource might not be specified correctly, or its status might not be ACTIVE.

RestoreTableFromBackupError

Error type for the RestoreTableFromBackup operation.

RestoreTableToPointInTimeError

Error type for the RestoreTableToPointInTime operation.

ScanError

Error type for the Scan operation.

TableAlreadyExistsException

A target table with the specified name already exists.

TableInUseException

A target table with the specified name is either being created or deleted.

TableNotFoundException

A source table with the name TableName does not currently exist within the subscriber's account or the subscriber is operating in the wrong Amazon Web Services Region.

TagResourceError

Error type for the TagResource operation.

TransactGetItemsError

Error type for the TransactGetItems operation.

TransactWriteItemsError

Error type for the TransactWriteItems operation.

TransactionCanceledException

The entire transaction request was canceled.

TransactionConflictException

Operation was rejected because there is an ongoing transaction for the item.

TransactionInProgressException

The transaction with the given request token is already in progress.

UntagResourceError

Error type for the UntagResource operation.

UpdateContinuousBackupsError

Error type for the UpdateContinuousBackups operation.

UpdateContributorInsightsError

Error type for the UpdateContributorInsights operation.

UpdateGlobalTableError

Error type for the UpdateGlobalTable operation.

UpdateGlobalTableSettingsError

Error type for the UpdateGlobalTableSettings operation.

UpdateItemError

Error type for the UpdateItem operation.

UpdateTableError

Error type for the UpdateTable operation.

UpdateTableReplicaAutoScalingError

Error type for the UpdateTableReplicaAutoScaling operation.

UpdateTimeToLiveError

Error type for the UpdateTimeToLive operation.

Enums

BatchExecuteStatementErrorKind

Types of errors that can occur for the BatchExecuteStatement operation.

BatchGetItemErrorKind

Types of errors that can occur for the BatchGetItem operation.

BatchWriteItemErrorKind

Types of errors that can occur for the BatchWriteItem operation.

CreateBackupErrorKind

Types of errors that can occur for the CreateBackup operation.

CreateGlobalTableErrorKind

Types of errors that can occur for the CreateGlobalTable operation.

CreateTableErrorKind

Types of errors that can occur for the CreateTable operation.

DeleteBackupErrorKind

Types of errors that can occur for the DeleteBackup operation.

DeleteItemErrorKind

Types of errors that can occur for the DeleteItem operation.

DeleteTableErrorKind

Types of errors that can occur for the DeleteTable operation.

DescribeBackupErrorKind

Types of errors that can occur for the DescribeBackup operation.

DescribeContinuousBackupsErrorKind

Types of errors that can occur for the DescribeContinuousBackups operation.

DescribeContributorInsightsErrorKind

Types of errors that can occur for the DescribeContributorInsights operation.

DescribeEndpointsErrorKind

Types of errors that can occur for the DescribeEndpoints operation.

DescribeExportErrorKind

Types of errors that can occur for the DescribeExport operation.

DescribeGlobalTableErrorKind

Types of errors that can occur for the DescribeGlobalTable operation.

DescribeGlobalTableSettingsErrorKind

Types of errors that can occur for the DescribeGlobalTableSettings operation.

DescribeKinesisStreamingDestinationErrorKind

Types of errors that can occur for the DescribeKinesisStreamingDestination operation.

DescribeLimitsErrorKind

Types of errors that can occur for the DescribeLimits operation.

DescribeTableErrorKind

Types of errors that can occur for the DescribeTable operation.

DescribeTableReplicaAutoScalingErrorKind

Types of errors that can occur for the DescribeTableReplicaAutoScaling operation.

DescribeTimeToLiveErrorKind

Types of errors that can occur for the DescribeTimeToLive operation.

DisableKinesisStreamingDestinationErrorKind

Types of errors that can occur for the DisableKinesisStreamingDestination operation.

EnableKinesisStreamingDestinationErrorKind

Types of errors that can occur for the EnableKinesisStreamingDestination operation.

ExecuteStatementErrorKind

Types of errors that can occur for the ExecuteStatement operation.

ExecuteTransactionErrorKind

Types of errors that can occur for the ExecuteTransaction operation.

ExportTableToPointInTimeErrorKind

Types of errors that can occur for the ExportTableToPointInTime operation.

GetItemErrorKind

Types of errors that can occur for the GetItem operation.

ListBackupsErrorKind

Types of errors that can occur for the ListBackups operation.

ListContributorInsightsErrorKind

Types of errors that can occur for the ListContributorInsights operation.

ListExportsErrorKind

Types of errors that can occur for the ListExports operation.

ListGlobalTablesErrorKind

Types of errors that can occur for the ListGlobalTables operation.

ListTablesErrorKind

Types of errors that can occur for the ListTables operation.

ListTagsOfResourceErrorKind

Types of errors that can occur for the ListTagsOfResource operation.

PutItemErrorKind

Types of errors that can occur for the PutItem operation.

QueryErrorKind

Types of errors that can occur for the Query operation.

RestoreTableFromBackupErrorKind

Types of errors that can occur for the RestoreTableFromBackup operation.

RestoreTableToPointInTimeErrorKind

Types of errors that can occur for the RestoreTableToPointInTime operation.

ScanErrorKind

Types of errors that can occur for the Scan operation.

TagResourceErrorKind

Types of errors that can occur for the TagResource operation.

TransactGetItemsErrorKind

Types of errors that can occur for the TransactGetItems operation.

TransactWriteItemsErrorKind

Types of errors that can occur for the TransactWriteItems operation.

UntagResourceErrorKind

Types of errors that can occur for the UntagResource operation.

UpdateContinuousBackupsErrorKind

Types of errors that can occur for the UpdateContinuousBackups operation.

UpdateContributorInsightsErrorKind

Types of errors that can occur for the UpdateContributorInsights operation.

UpdateGlobalTableErrorKind

Types of errors that can occur for the UpdateGlobalTable operation.

UpdateGlobalTableSettingsErrorKind

Types of errors that can occur for the UpdateGlobalTableSettings operation.

UpdateItemErrorKind

Types of errors that can occur for the UpdateItem operation.

UpdateTableErrorKind

Types of errors that can occur for the UpdateTable operation.

UpdateTableReplicaAutoScalingErrorKind

Types of errors that can occur for the UpdateTableReplicaAutoScaling operation.

UpdateTimeToLiveErrorKind

Types of errors that can occur for the UpdateTimeToLive operation.