Struct aws_sdk_dynamodb::input::create_table_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for CreateTableInput

Implementations

impl Builder

pub fn attribute_definitions(
self,
input: impl Into<AttributeDefinition>
) -> Self

Appends an item to attribute_definitions.

To override the contents of this collection use set_attribute_definitions.

An array of attributes that describe the key schema for the table and indexes.

pub fn set_attribute_definitions(
self,
input: Option<Vec<AttributeDefinition>>
) -> Self

An array of attributes that describe the key schema for the table and indexes.

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table to create.

pub fn set_table_name(self, input: Option<String>) -> Self

The name of the table to create.

pub fn key_schema(self, input: impl Into<KeySchemaElement>) -> Self

Appends an item to key_schema.

To override the contents of this collection use set_key_schema.

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

pub fn set_key_schema(self, input: Option<Vec<KeySchemaElement>>) -> Self

Specifies the attributes that make up the primary key for a table or an index. The attributes in KeySchema must also be defined in the AttributeDefinitions array. For more information, see Data Model in the Amazon DynamoDB Developer Guide.

Each KeySchemaElement in the array is composed of:

AttributeName - The name of this key attribute.

KeyType - The role that the key attribute will assume:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from the DynamoDB usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For a simple primary key (partition key), you must provide exactly one element with a KeyType of HASH.

For a composite primary key (partition key and sort key), you must provide exactly two elements, in this order: The first element must have a KeyType of HASH, and the second element must have a KeyType of RANGE.

For more information, see Working with Tables in the Amazon DynamoDB Developer Guide.

pub fn local_secondary_indexes(
self,
input: impl Into<LocalSecondaryIndex>
) -> Self

Appends an item to local_secondary_indexes.

To override the contents of this collection use set_local_secondary_indexes.

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

pub fn set_local_secondary_indexes(
self,
input: Option<Vec<LocalSecondaryIndex>>
) -> Self

One or more local secondary indexes (the maximum is 5) to be created on the table. Each index is scoped to a given partition key value. There is a 10 GB size limit per partition key value; otherwise, the size of a local secondary index is unconstrained.

Each local secondary index in the array includes the following:

IndexName - The name of the local secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the local secondary index. The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

pub fn global_secondary_indexes(
self,
input: impl Into<GlobalSecondaryIndex>
) -> Self

Appends an item to global_secondary_indexes.

To override the contents of this collection use set_global_secondary_indexes.

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

pub fn set_global_secondary_indexes(
self,
input: Option<Vec<GlobalSecondaryIndex>>
) -> Self

One or more global secondary indexes (the maximum is 20) to be created on the table. Each global secondary index in the array includes the following:

IndexName - The name of the global secondary index. Must be unique only for this table.

KeySchema - Specifies the key schema for the global secondary index.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units.

pub fn billing_mode(self, input: BillingMode) -> Self

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

pub fn set_billing_mode(self, input: Option<BillingMode>) -> Self

Controls how you are charged for read and write throughput and how you manage capacity. This setting can be changed later.

PROVISIONED - We recommend using PROVISIONED for predictable workloads. PROVISIONED sets the billing mode to Provisioned Mode.

PAY_PER_REQUEST - We recommend using PAY_PER_REQUEST for unpredictable workloads. PAY_PER_REQUEST sets the billing mode to On-Demand Mode.

pub fn provisioned_throughput(self, input: ProvisionedThroughput) -> Self

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

pub fn set_provisioned_throughput(
self,
input: Option<ProvisionedThroughput>
) -> Self

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

If you set BillingMode as PROVISIONED, you must specify this property. If you set BillingMode as PAY_PER_REQUEST, you cannot specify this property.

For current minimum and maximum provisioned throughput values, see Service, Account, and Table Quotas in the Amazon DynamoDB Developer Guide.

pub fn stream_specification(self, input: StreamSpecification) -> Self

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

pub fn set_stream_specification(
self,
input: Option<StreamSpecification>
) -> Self

The settings for DynamoDB Streams on the table. These settings consist of:

StreamEnabled - Indicates whether DynamoDB Streams is to be enabled (true) or disabled (false).

StreamViewType - When an item in the table is modified, StreamViewType determines what information is written to the table's stream. Valid values for StreamViewType are:

KEYS_ONLY - Only the key attributes of the modified item are written to the stream.

NEW_IMAGE - The entire item, as it appears after it was modified, is written to the stream.

OLD_IMAGE - The entire item, as it appeared before it was modified, is written to the stream.

NEW_AND_OLD_IMAGES - Both the new and the old item images of the item are written to the stream.

pub fn sse_specification(self, input: SseSpecification) -> Self

Represents the settings used to enable server-side encryption.

pub fn set_sse_specification(self, input: Option<SseSpecification>) -> Self

Represents the settings used to enable server-side encryption.

pub fn tags(self, input: impl Into<Tag>) -> Self

Appends an item to tags.

To override the contents of this collection use set_tags.

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

A list of key-value pairs to label the table. For more information, see Tagging for DynamoDB.

pub fn build(self) -> Result<CreateTableInput, BuildError>

Consumes the builder and constructs a CreateTableInput