Module aws_sdk_dynamodb::model

Data structures used by operation inputs/outputs.

Modules

archival_summary

See ArchivalSummary

attribute_definition

See AttributeDefinition

attribute_value_update

See AttributeValueUpdate

auto_scaling_policy_description

See AutoScalingPolicyDescription

auto_scaling_policy_update

See AutoScalingPolicyUpdate

auto_scaling_settings_description

See AutoScalingSettingsDescription

auto_scaling_settings_update

See AutoScalingSettingsUpdate

auto_scaling_target_tracking_scaling_policy_configuration_description

See AutoScalingTargetTrackingScalingPolicyConfigurationDescription

auto_scaling_target_tracking_scaling_policy_configuration_update

See AutoScalingTargetTrackingScalingPolicyConfigurationUpdate

backup_description

See BackupDescription

backup_details

See BackupDetails

backup_summary

See BackupSummary

batch_statement_error

See BatchStatementError

batch_statement_request

See BatchStatementRequest

batch_statement_response

See BatchStatementResponse

billing_mode_summary

See BillingModeSummary

cancellation_reason

See CancellationReason

capacity

See Capacity

condition

See Condition

condition_check

See ConditionCheck

consumed_capacity

See ConsumedCapacity

continuous_backups_description

See ContinuousBackupsDescription

contributor_insights_summary

See ContributorInsightsSummary

create_global_secondary_index_action

See CreateGlobalSecondaryIndexAction

create_replica_action

See CreateReplicaAction

create_replication_group_member_action

See CreateReplicationGroupMemberAction

delete

See Delete

delete_global_secondary_index_action

See DeleteGlobalSecondaryIndexAction

delete_replica_action

See DeleteReplicaAction

delete_replication_group_member_action

See DeleteReplicationGroupMemberAction

delete_request

See DeleteRequest

endpoint

See Endpoint

expected_attribute_value

See ExpectedAttributeValue

export_description

See ExportDescription

export_summary

See ExportSummary

failure_exception

See FailureException

get

See Get

global_secondary_index

See GlobalSecondaryIndex

global_secondary_index_auto_scaling_update

See GlobalSecondaryIndexAutoScalingUpdate

global_secondary_index_description

See GlobalSecondaryIndexDescription

global_secondary_index_info

See GlobalSecondaryIndexInfo

global_secondary_index_update

See GlobalSecondaryIndexUpdate

global_table

See GlobalTable

global_table_description

See GlobalTableDescription

global_table_global_secondary_index_settings_update

See GlobalTableGlobalSecondaryIndexSettingsUpdate

item_collection_metrics

See ItemCollectionMetrics

item_response

See ItemResponse

key_schema_element

See KeySchemaElement

keys_and_attributes

See KeysAndAttributes

kinesis_data_stream_destination

See KinesisDataStreamDestination

local_secondary_index

See LocalSecondaryIndex

local_secondary_index_description

See LocalSecondaryIndexDescription

local_secondary_index_info

See LocalSecondaryIndexInfo

parameterized_statement

See ParameterizedStatement

point_in_time_recovery_description

See PointInTimeRecoveryDescription

point_in_time_recovery_specification

See PointInTimeRecoverySpecification

projection

See Projection

provisioned_throughput

See ProvisionedThroughput

provisioned_throughput_description

See ProvisionedThroughputDescription

provisioned_throughput_override

See ProvisionedThroughputOverride

put

See Put

put_request

See PutRequest

replica

See Replica

replica_auto_scaling_description

See ReplicaAutoScalingDescription

replica_auto_scaling_update

See ReplicaAutoScalingUpdate

replica_description

See ReplicaDescription

replica_global_secondary_index

See ReplicaGlobalSecondaryIndex

replica_global_secondary_index_auto_scaling_description

See ReplicaGlobalSecondaryIndexAutoScalingDescription

replica_global_secondary_index_auto_scaling_update

See ReplicaGlobalSecondaryIndexAutoScalingUpdate

replica_global_secondary_index_description

See ReplicaGlobalSecondaryIndexDescription

replica_global_secondary_index_settings_description

See ReplicaGlobalSecondaryIndexSettingsDescription

replica_global_secondary_index_settings_update

See ReplicaGlobalSecondaryIndexSettingsUpdate

replica_settings_description

See ReplicaSettingsDescription

replica_settings_update

See ReplicaSettingsUpdate

replica_update

See ReplicaUpdate

replication_group_update

See ReplicationGroupUpdate

restore_summary

See RestoreSummary

source_table_details

See SourceTableDetails

source_table_feature_details

See SourceTableFeatureDetails

sse_description

See SseDescription

sse_specification

See SseSpecification

stream_specification

See StreamSpecification

table_auto_scaling_description

See TableAutoScalingDescription

table_class_summary

See TableClassSummary

table_description

See TableDescription

tag

See Tag

time_to_live_description

See TimeToLiveDescription

time_to_live_specification

See TimeToLiveSpecification

transact_get_item

See TransactGetItem

transact_write_item

See TransactWriteItem

update

See Update

update_global_secondary_index_action

See UpdateGlobalSecondaryIndexAction

update_replication_group_member_action

See UpdateReplicationGroupMemberAction

write_request

See WriteRequest

Structs

ArchivalSummary

Contains details of a table archival operation.

AttributeDefinition

Represents an attribute for describing the key schema for the table and indexes.

AttributeValueUpdate

For the UpdateItem operation, represents the attributes to be modified, the action to perform on each, and the new value for each.

AutoScalingPolicyDescription

Represents the properties of the scaling policy.

AutoScalingPolicyUpdate

Represents the auto scaling policy to be modified.

AutoScalingSettingsDescription

Represents the auto scaling settings for a global table or global secondary index.

AutoScalingSettingsUpdate

Represents the auto scaling settings to be modified for a global table or global secondary index.

AutoScalingTargetTrackingScalingPolicyConfigurationDescription

Represents the properties of a target tracking scaling policy.

AutoScalingTargetTrackingScalingPolicyConfigurationUpdate

Represents the settings of a target tracking scaling policy that will be modified.

BackupDescription

Contains the description of the backup created for the table.

BackupDetails

Contains the details of the backup created for the table.

BackupSummary

Contains details for the backup.

BatchStatementError

An error associated with a statement in a PartiQL batch that was run.

BatchStatementRequest

A PartiQL batch statement request.

BatchStatementResponse

A PartiQL batch statement response..

BillingModeSummary

Contains the details for the read/write capacity mode.

CancellationReason

An ordered list of errors for each item in the request which caused the transaction to get cancelled. The values of the list are ordered according to the ordering of the TransactWriteItems request parameter. If no error occurred for the associated item an error with a Null code and Null message will be present.

Capacity

Represents the amount of provisioned throughput capacity consumed on a table or an index.

Condition

Represents the selection criteria for a Query or Scan operation:

ConditionCheck

Represents a request to perform a check that an item exists or to check the condition of specific attributes of the item.

ConsumedCapacity

The capacity units consumed by an operation. The data returned includes the total provisioned throughput consumed, along with statistics for the table and any indexes involved in the operation. ConsumedCapacity is only returned if the request asked for it. For more information, see Provisioned Throughput in the Amazon DynamoDB Developer Guide.

ContinuousBackupsDescription

Represents the continuous backups and point in time recovery settings on the table.

ContributorInsightsSummary

Represents a Contributor Insights summary entry.

CreateGlobalSecondaryIndexAction

Represents a new global secondary index to be added to an existing table.

CreateReplicaAction

Represents a replica to be added.

CreateReplicationGroupMemberAction

Represents a replica to be created.

Delete

Represents a request to perform a DeleteItem operation.

DeleteGlobalSecondaryIndexAction

Represents a global secondary index to be deleted from an existing table.

DeleteReplicaAction

Represents a replica to be removed.

DeleteReplicationGroupMemberAction

Represents a replica to be deleted.

DeleteRequest

Represents a request to perform a DeleteItem operation on an item.

Endpoint

An endpoint information details.

ExpectedAttributeValue

Represents a condition to be compared with an attribute value. This condition can be used with DeleteItem, PutItem, or UpdateItem operations; if the comparison evaluates to true, the operation succeeds; if not, the operation fails. You can use ExpectedAttributeValue in one of two different ways:

ExportDescription

Represents the properties of the exported table.

ExportSummary

Summary information about an export task.

FailureException

Represents a failure a contributor insights operation.

Get

Specifies an item and related attribute values to retrieve in a TransactGetItem object.

GlobalSecondaryIndex

Represents the properties of a global secondary index.

GlobalSecondaryIndexAutoScalingUpdate

Represents the auto scaling settings of a global secondary index for a global table that will be modified.

GlobalSecondaryIndexDescription

Represents the properties of a global secondary index.

GlobalSecondaryIndexInfo

Represents the properties of a global secondary index for the table when the backup was created.

GlobalSecondaryIndexUpdate

Represents one of the following:

GlobalTable

Represents the properties of a global table.

GlobalTableDescription

Contains details about the global table.

GlobalTableGlobalSecondaryIndexSettingsUpdate

Represents the settings of a global secondary index for a global table that will be modified.

ItemCollectionMetrics

Information about item collections, if any, that were affected by the operation. ItemCollectionMetrics is only returned if the request asked for it. If the table does not have any local secondary indexes, this information is not returned in the response.

ItemResponse

Details for the requested item.

KeySchemaElement

Represents a single element of a key schema. A key schema specifies the attributes that make up the primary key of a table, or the key attributes of an index.

KeysAndAttributes

Represents a set of primary keys and, for each key, the attributes to retrieve from the table.

KinesisDataStreamDestination

Describes a Kinesis data stream destination.

LocalSecondaryIndex

Represents the properties of a local secondary index.

LocalSecondaryIndexDescription

Represents the properties of a local secondary index.

LocalSecondaryIndexInfo

Represents the properties of a local secondary index for the table when the backup was created.

ParameterizedStatement

Represents a PartiQL statment that uses parameters.

PointInTimeRecoveryDescription

The description of the point in time settings applied to the table.

PointInTimeRecoverySpecification

Represents the settings used to enable point in time recovery.

Projection

Represents attributes that are copied (projected) from the table into an index. These are in addition to the primary key attributes and index key attributes, which are automatically projected.

ProvisionedThroughput

Represents the provisioned throughput settings for a specified table or index. The settings can be modified using the UpdateTable operation.

ProvisionedThroughputDescription

Represents the provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

ProvisionedThroughputOverride

Replica-specific provisioned throughput settings. If not specified, uses the source table's provisioned throughput settings.

Put

Represents a request to perform a PutItem operation.

PutRequest

Represents a request to perform a PutItem operation on an item.

Replica

Represents the properties of a replica.

ReplicaAutoScalingDescription

Represents the auto scaling settings of the replica.

ReplicaAutoScalingUpdate

Represents the auto scaling settings of a replica that will be modified.

ReplicaDescription

Contains the details of the replica.

ReplicaGlobalSecondaryIndex

Represents the properties of a replica global secondary index.

ReplicaGlobalSecondaryIndexAutoScalingDescription

Represents the auto scaling configuration for a replica global secondary index.

ReplicaGlobalSecondaryIndexAutoScalingUpdate

Represents the auto scaling settings of a global secondary index for a replica that will be modified.

ReplicaGlobalSecondaryIndexDescription

Represents the properties of a replica global secondary index.

ReplicaGlobalSecondaryIndexSettingsDescription

Represents the properties of a global secondary index.

ReplicaGlobalSecondaryIndexSettingsUpdate

Represents the settings of a global secondary index for a global table that will be modified.

ReplicaSettingsDescription

Represents the properties of a replica.

ReplicaSettingsUpdate

Represents the settings for a global table in a Region that will be modified.

ReplicaUpdate

Represents one of the following:

ReplicationGroupUpdate

Represents one of the following:

RestoreSummary

Contains details for the restore.

SourceTableDetails

Contains the details of the table when the backup was created.

SourceTableFeatureDetails

Contains the details of the features enabled on the table when the backup was created. For example, LSIs, GSIs, streams, TTL.

SseDescription

The description of the server-side encryption status on the specified table.

SseSpecification

Represents the settings used to enable server-side encryption.

StreamSpecification

Represents the DynamoDB Streams configuration for a table in DynamoDB.

TableAutoScalingDescription

Represents the auto scaling configuration for a global table.

TableClassSummary

Contains details of the table class.

TableDescription

Represents the properties of a table.

Tag

Describes a tag. A tag is a key-value pair. You can add up to 50 tags to a single DynamoDB table.

TimeToLiveDescription

The description of the Time to Live (TTL) status on the specified table.

TimeToLiveSpecification

Represents the settings used to enable or disable Time to Live (TTL) for the specified table.

TransactGetItem

Specifies an item to be retrieved as part of the transaction.

TransactWriteItem

A list of requests that can perform update, put, delete, or check operations on multiple items in one or more tables atomically.

Update

Represents a request to perform an UpdateItem operation.

UpdateGlobalSecondaryIndexAction

Represents the new provisioned throughput settings to be applied to a global secondary index.

UpdateReplicationGroupMemberAction

Represents a replica to be modified.

WriteRequest

Represents an operation to perform - either DeleteItem or PutItem. You can only request one of these operations, not both, in a single WriteRequest. If you do need to perform both of these operations, you need to provide two separate WriteRequest objects.

Enums

AttributeAction

AttributeValue

Represents the data for an attribute.

BackupStatus

BackupType

BackupTypeFilter

BatchStatementErrorCodeEnum

BillingMode

ComparisonOperator

ConditionalOperator

ContinuousBackupsStatus

ContributorInsightsAction

ContributorInsightsStatus

DestinationStatus

ExportFormat

ExportStatus

GlobalTableStatus

IndexStatus

KeyType

PointInTimeRecoveryStatus

ProjectionType

ReplicaStatus

ReturnConsumedCapacity

Determines the level of detail about either provisioned or on-demand throughput

ReturnItemCollectionMetrics

ReturnValue

ReturnValuesOnConditionCheckFailure

S3SseAlgorithm

ScalarAttributeType

Select

SseStatus

SseType

StreamViewType

TableClass

TableStatus

TimeToLiveStatus