Struct aws_sdk_dynamodb::model::table_description::Builder

#[non_exhaustive]
pub struct Builder { /* private fields */ }

A builder for TableDescription

Implementations

impl Builder

pub fn attribute_definitions(self, input: AttributeDefinition) -> Self

Appends an item to attribute_definitions.

To override the contents of this collection use set_attribute_definitions.

An array of AttributeDefinition objects. Each of these objects describes one attribute in the table and index key schema.

Each AttributeDefinition object in this array is composed of:

AttributeName - The name of the attribute.

AttributeType - The data type for the attribute.

pub fn set_attribute_definitions(
self,
input: Option<Vec<AttributeDefinition>>
) -> Self

An array of AttributeDefinition objects. Each of these objects describes one attribute in the table and index key schema.

Each AttributeDefinition object in this array is composed of:

AttributeName - The name of the attribute.

AttributeType - The data type for the attribute.

pub fn table_name(self, input: impl Into<String>) -> Self

The name of the table.

pub fn set_table_name(self, input: Option<String>) -> Self

The name of the table.

pub fn key_schema(self, input: KeySchemaElement) -> Self

Appends an item to key_schema.

To override the contents of this collection use set_key_schema.

The primary key structure for the table. Each KeySchemaElement consists of:

AttributeName - The name of the attribute.

KeyType - The role of the attribute:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For more information about primary keys, see Primary Key in the Amazon DynamoDB Developer Guide.

pub fn set_key_schema(self, input: Option<Vec<KeySchemaElement>>) -> Self

The primary key structure for the table. Each KeySchemaElement consists of:

AttributeName - The name of the attribute.

KeyType - The role of the attribute:

HASH - partition key

RANGE - sort key

The partition key of an item is also known as its hash attribute. The term "hash attribute" derives from DynamoDB's usage of an internal hash function to evenly distribute data items across partitions, based on their partition key values.

The sort key of an item is also known as its range attribute. The term "range attribute" derives from the way DynamoDB stores items with the same partition key physically close together, in sorted order by the sort key value.

For more information about primary keys, see Primary Key in the Amazon DynamoDB Developer Guide.

pub fn table_status(self, input: TableStatus) -> Self

The current state of the table:

CREATING - The table is being created.

UPDATING - The table is being updated.

DELETING - The table is being deleted.

ACTIVE - The table is ready for use.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The KMS key used to encrypt the table in inaccessible. Table operations may fail due to failure to use the KMS key. DynamoDB will initiate the table archival process when a table's KMS key remains inaccessible for more than seven days.

ARCHIVING - The table is being archived. Operations are not allowed until archival is complete.

ARCHIVED - The table has been archived. See the ArchivalReason for more information.

pub fn set_table_status(self, input: Option<TableStatus>) -> Self

The current state of the table:

CREATING - The table is being created.

UPDATING - The table is being updated.

DELETING - The table is being deleted.

ACTIVE - The table is ready for use.

INACCESSIBLE_ENCRYPTION_CREDENTIALS - The KMS key used to encrypt the table in inaccessible. Table operations may fail due to failure to use the KMS key. DynamoDB will initiate the table archival process when a table's KMS key remains inaccessible for more than seven days.

ARCHIVING - The table is being archived. Operations are not allowed until archival is complete.

ARCHIVED - The table has been archived. See the ArchivalReason for more information.

pub fn creation_date_time(self, input: DateTime) -> Self

The date and time when the table was created, in UNIX epoch time format.

pub fn set_creation_date_time(self, input: Option<DateTime>) -> Self

The date and time when the table was created, in UNIX epoch time format.

pub fn provisioned_throughput(
self,
input: ProvisionedThroughputDescription
) -> Self

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

pub fn set_provisioned_throughput(
self,
input: Option<ProvisionedThroughputDescription>
) -> Self

The provisioned throughput settings for the table, consisting of read and write capacity units, along with data about increases and decreases.

pub fn table_size_bytes(self, input: i64) -> Self

The total size of the specified table, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn set_table_size_bytes(self, input: Option<i64>) -> Self

The total size of the specified table, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn item_count(self, input: i64) -> Self

The number of items in the specified table. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn set_item_count(self, input: Option<i64>) -> Self

The number of items in the specified table. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

pub fn table_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the table.

pub fn set_table_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the table.

pub fn table_id(self, input: impl Into<String>) -> Self

Unique identifier for the table for which the backup was created.

pub fn set_table_id(self, input: Option<String>) -> Self

Unique identifier for the table for which the backup was created.

pub fn billing_mode_summary(self, input: BillingModeSummary) -> Self

Contains the details for the read/write capacity mode.

pub fn set_billing_mode_summary(self, input: Option<BillingModeSummary>) -> Self

Contains the details for the read/write capacity mode.

pub fn local_secondary_indexes(
self,
input: LocalSecondaryIndexDescription
) -> Self

Appends an item to local_secondary_indexes.

To override the contents of this collection use set_local_secondary_indexes.

Represents one or more local secondary indexes on the table. Each index is scoped to a given partition key value. Tables with one or more local secondary indexes are subject to an item collection size limit, where the amount of data within a given item collection cannot exceed 10 GB. Each element is composed of:

IndexName - The name of the local secondary index.

KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

IndexSizeBytes - Represents the total size of the index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

ItemCount - Represents the number of items in the index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

If the table is in the DELETING state, no information about indexes will be returned.

pub fn set_local_secondary_indexes(
self,
input: Option<Vec<LocalSecondaryIndexDescription>>
) -> Self

Represents one or more local secondary indexes on the table. Each index is scoped to a given partition key value. Tables with one or more local secondary indexes are subject to an item collection size limit, where the amount of data within a given item collection cannot exceed 10 GB. Each element is composed of:

IndexName - The name of the local secondary index.

KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - Only the specified table attributes are projected into the index. The list of projected attributes is in NonKeyAttributes.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

IndexSizeBytes - Represents the total size of the index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

ItemCount - Represents the number of items in the index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

If the table is in the DELETING state, no information about indexes will be returned.

pub fn global_secondary_indexes(
self,
input: GlobalSecondaryIndexDescription
) -> Self

Appends an item to global_secondary_indexes.

To override the contents of this collection use set_global_secondary_indexes.

The global secondary indexes, if any, on the table. Each index is scoped to a given partition key value. Each element is composed of:

Backfilling - If true, then the index is currently in the backfilling phase. Backfilling occurs only when a new global secondary index is added to the table. It is the process by which DynamoDB populates the new index with data from the table. (This attribute does not appear for indexes that were created during a CreateTable operation.)

You can delete an index that is being created during the Backfilling phase when IndexStatus is set to CREATING and Backfilling is true. You can't delete the index that is being created when IndexStatus is set to CREATING and Backfilling is false. (This attribute does not appear for indexes that were created during a CreateTable operation.)

IndexName - The name of the global secondary index.

IndexSizeBytes - The total size of the global secondary index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

IndexStatus - The current status of the global secondary index:

CREATING - The index is being created.

UPDATING - The index is being updated.

DELETING - The index is being deleted.

ACTIVE - The index is ready for use.

ItemCount - The number of items in the global secondary index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - In addition to the attributes described in KEYS_ONLY, the secondary index will include other non-key attributes that you specify.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units, along with data about increases and decreases.

If the table is in the DELETING state, no information about indexes will be returned.

pub fn set_global_secondary_indexes(
self,
input: Option<Vec<GlobalSecondaryIndexDescription>>
) -> Self

The global secondary indexes, if any, on the table. Each index is scoped to a given partition key value. Each element is composed of:

Backfilling - If true, then the index is currently in the backfilling phase. Backfilling occurs only when a new global secondary index is added to the table. It is the process by which DynamoDB populates the new index with data from the table. (This attribute does not appear for indexes that were created during a CreateTable operation.)

You can delete an index that is being created during the Backfilling phase when IndexStatus is set to CREATING and Backfilling is true. You can't delete the index that is being created when IndexStatus is set to CREATING and Backfilling is false. (This attribute does not appear for indexes that were created during a CreateTable operation.)

IndexName - The name of the global secondary index.

IndexSizeBytes - The total size of the global secondary index, in bytes. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

IndexStatus - The current status of the global secondary index:

CREATING - The index is being created.

UPDATING - The index is being updated.

DELETING - The index is being deleted.

ACTIVE - The index is ready for use.

ItemCount - The number of items in the global secondary index. DynamoDB updates this value approximately every six hours. Recent changes might not be reflected in this value.

KeySchema - Specifies the complete index key schema. The attribute names in the key schema must be between 1 and 255 characters (inclusive). The key schema must begin with the same partition key as the table.

Projection - Specifies attributes that are copied (projected) from the table into the index. These are in addition to the primary key attributes and index key attributes, which are automatically projected. Each attribute specification is composed of:

ProjectionType - One of the following:

KEYS_ONLY - Only the index and primary keys are projected into the index.

INCLUDE - In addition to the attributes described in KEYS_ONLY, the secondary index will include other non-key attributes that you specify.

ALL - All of the table attributes are projected into the index.

NonKeyAttributes - A list of one or more non-key attribute names that are projected into the secondary index. The total count of attributes provided in NonKeyAttributes, summed across all of the secondary indexes, must not exceed 100. If you project the same attribute into two different indexes, this counts as two distinct attributes when determining the total.

ProvisionedThroughput - The provisioned throughput settings for the global secondary index, consisting of read and write capacity units, along with data about increases and decreases.

If the table is in the DELETING state, no information about indexes will be returned.

pub fn stream_specification(self, input: StreamSpecification) -> Self

The current DynamoDB Streams configuration for the table.

pub fn set_stream_specification(self, input: Option<StreamSpecification>) -> Self

The current DynamoDB Streams configuration for the table.

pub fn latest_stream_label(self, input: impl Into<String>) -> Self

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

Amazon Web Services customer ID

Table name

StreamLabel

pub fn set_latest_stream_label(self, input: Option<String>) -> Self

A timestamp, in ISO 8601 format, for this stream.

Note that LatestStreamLabel is not a unique identifier for the stream, because it is possible that a stream from another table might have the same timestamp. However, the combination of the following three elements is guaranteed to be unique:

Amazon Web Services customer ID

Table name

StreamLabel

pub fn latest_stream_arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the latest stream for this table.

pub fn set_latest_stream_arn(self, input: Option<String>) -> Self

The Amazon Resource Name (ARN) that uniquely identifies the latest stream for this table.

pub fn global_table_version(self, input: impl Into<String>) -> Self

Represents the version of global tables in use, if the table is replicated across Amazon Web Services Regions.

pub fn set_global_table_version(self, input: Option<String>) -> Self

Represents the version of global tables in use, if the table is replicated across Amazon Web Services Regions.

pub fn replicas(self, input: ReplicaDescription) -> Self

Appends an item to replicas.

To override the contents of this collection use set_replicas.

Represents replicas of the table.

pub fn set_replicas(self, input: Option<Vec<ReplicaDescription>>) -> Self

Represents replicas of the table.

pub fn restore_summary(self, input: RestoreSummary) -> Self

Contains details for the restore.

pub fn set_restore_summary(self, input: Option<RestoreSummary>) -> Self

Contains details for the restore.

pub fn sse_description(self, input: SseDescription) -> Self

The description of the server-side encryption status on the specified table.

pub fn set_sse_description(self, input: Option<SseDescription>) -> Self

The description of the server-side encryption status on the specified table.

pub fn archival_summary(self, input: ArchivalSummary) -> Self

Contains information about the table archive.

pub fn set_archival_summary(self, input: Option<ArchivalSummary>) -> Self

Contains information about the table archive.

pub fn table_class_summary(self, input: TableClassSummary) -> Self

Contains details of the table class.

pub fn set_table_class_summary(self, input: Option<TableClassSummary>) -> Self

Contains details of the table class.

pub fn build(self) -> TableDescription

Consumes the builder and constructs a TableDescription